% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_higherorder.R
\name{add_higherorder}
\alias{add_higherorder}
\title{Add multi-nucleotide information to a motif.}
\usage{
add_higherorder(motif, sequences, add.k = 2:3, threshold = 0.01,
  RC = FALSE, motifs.perseq = 1, BPPARAM = bpparam())
}
\arguments{
\item{motif}{universalmotif object.}

\item{sequences}{DNAStringSet.}

\item{add.k}{Numeric.}

\item{threshold}{Numeric.}

\item{RC}{Logical.}

\item{motifs.perseq}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\description{
Add multi-nucleotide information to a motif.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
