% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hmm.R
\name{add_hmm}
\alias{add_hmm}
\title{Add first and/or second order HMM data to a motif.}
\usage{
add_hmm(motif, sequences, add.first = TRUE, add.second = TRUE,
  threshold = 0.01, RC = FALSE, motifs.perseq = 1, BPPARAM = bpparam())
}
\arguments{
\item{motif}{universalmotif object.}

\item{sequences}{DNAStringSet.}

\item{add.first}{Logical.}

\item{add.second}{Logical.}

\item{threshold}{Numeric.}

\item{RC}{Logical.}

\item{motifs.perseq}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\description{
Add first and/or second order HMM data to a motif.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
