% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_motifs.R
\name{align_motifs}
\alias{align_motifs}
\title{Align motifs using msaClustalW}
\usage{
align_motifs(motifs, cluster = "default", substitutionMatrix = "iub", ...)
}
\arguments{
\item{motifs}{List of motifs}

\item{cluster}{nj (default) or upgma}

\item{substitutionMatrix}{iub or clustalw}

\item{...}{Other settings for function msaClustalW}
}
\value{
List of motifs.
}
\description{
Align a list of motifs then return motifs of equal length.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
