% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArabidopsisPromoters}
\alias{ArabidopsisPromoters}
\title{Arabidopsis promoters.}
\format{\linkS4class{DNAStringSet}}
\usage{
ArabidopsisPromoters
}
\description{
50 Arabidopsis promoters, each 1000 bases long.
}
\keyword{datasets}
