% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sequences.R
\name{create_sequences}
\alias{create_sequences}
\title{Create random background sequences.}
\usage{
create_sequences(alphabet = "DNA", bkg, numseqs = 100, seqlen = 100,
  difreq, trifreq)
}
\arguments{
\item{alphabet}{Character. One of 'DNA', 'RNA', 'AA', or a string of letters
to be used as the alphabet.}

\item{bkg}{Numeric. Alphabet frequencies to use. If missing assumes uniform
frequencies. Not used if \code{difreq} or \code{trifreq} are
input.}

\item{numseqs}{Numeric. Number of sequences to generate.}

\item{seqlen}{Numeric. Length of random sequences.}

\item{difreq}{Numeric. Dinucleotide frequencies. DNA only. Must be a
named numeric vector of length 16.}

\item{trifreq}{Numeric. Trinucleotide frequencies. DNA only. Must be a 
named numeric vector of length 64.}
}
\value{
XStringSet object.
}
\description{
Create random background sequences.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
