% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tffm.R
\name{create_tffm}
\alias{create_tffm}
\title{Create a TFFM.}
\usage{
create_tffm(sequences, memefile, type = "first", ID = "Unknown",
  name = "Unknown", strand = "+", family = "Unknown", bkg = c(0.25,
  0.25, 0.25, 0.25), pseudocount = 1, BPPARAM = bpparam())
}
\arguments{
\item{sequences}{DNAStringSet (or list of). All sequences must be of the
same width.}

\item{memefile}{Character. File path to MEME output.}

\item{type}{Character. Currently only 'first' is supported.}

\item{ID}{Character. Motif ID.}

\item{name}{Character. Motif name.}

\item{strand}{Character. Motif strand.}

\item{family}{Character. Transcription factor family.}

\item{bkg}{Numeric. Background frequencies.}

\item{pseudocount}{Numeric. Pseudocount to be added when calculating TFFM.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
TFFMFirst object, TFFMDetail object, or a list of such objects.
}
\description{
If the sequences of all binding sites are available, a transcription
factor flexible model \insertCite{tffm}{universalmotif} can be generated.
TFFMs are based on hidden Markov models and are a more flexible
representation of a transcription factor binding site. The code to generate
TFFMs using this function is based on the python module 'TFFM'
\insertCite{tffm}{universalmotif}.
}
\details{
Code based on TFFM python module \insertCite{tffm}{universalmotif}.
   Currently the only support for this type of motif in Bioconductor is
   via the TFBSTools package \insertCite{tfbstools}{universalmotif}. As a
   result, the TFFM is stored using the TFFM class from TFBSTools.

   If a MEME output file is used to create the TFFM, then the 'strand'
   and 'bkg' fields will be taken from the MEME output. If there are
   multiple motifs in the output file, a TFFM will be created from each
   and a list of TFFM objects returned.
}
\examples{
   library(Biostrings)
   sites <- readDNAStringSet(system.file("extdata", "sites.txt",
                                         package = "universalmotif"))
   TFFM <- create_tffm(sites)
   TFBSTools::seqLogo(TFFM)

}
\references{
\insertRef{tffm}{universalmotif}
}
\seealso{
\code{\link{create_motif}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
