% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplemotif}
\alias{examplemotif}
\title{Example motif in \code{universalmotif} format.}
\format{\linkS4class{universalmotif}}
\usage{
examplemotif
}
\description{
A simple DNA motif. To recreate this motif:
\code{create_motif("TATAWAW", nsites = numeric())}
}
\keyword{datasets}
