% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fontDFroboto}
\alias{fontDFroboto}
\title{Polygon coordinates for plotting letters.}
\format{
\linkS4class{DataFrame}
}
\usage{
fontDFroboto
}
\description{
\code{DataFrame} of polygon coordinates used by \code{\link[=view_motifs]{view_motifs()}} for plotting
letters. It was generated using the \code{createPolygons} function from the
\code{gglogo} package for the font Roboto Medium.
}
\keyword{datasets}
