% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bkg.R
\name{get_bkg}
\alias{get_bkg}
\title{Calculate sequence background.}
\usage{
get_bkg(
  sequences,
  k = 1:3,
  as.prob = TRUE,
  pseudocount = 0,
  alphabet = NULL,
  to.meme = NULL,
  RC = FALSE,
  list.out = TRUE,
  nthreads = 1
)
}
\arguments{
\item{sequences}{\code{\link{XStringSet}} Input sequences. Note that if
multiple sequences are present, the results will be combined into one.}

\item{k}{\code{integer} Size of k-let. Background can be calculated for any
k-let size.}

\item{as.prob}{\code{logical(1)} Whether to return k-let counts or probabilities.}

\item{pseudocount}{\code{integer(1)} Add a count to each possible k-let. Prevents
any k-let from having 0 or 1 probabilities.}

\item{alphabet}{\code{character(1)} Provide a custom alphabet to calculate a
background for. If \code{NULL}, then standard letters will be assumed for
DNA, RNA and AA sequences, and all unique letters found will be used
for \code{BStringSet} type sequences.}

\item{to.meme}{If not \code{NULL}, then \code{\link[=get_bkg]{get_bkg()}} will return the sequence
background in MEME Markov Background Model format. Input for this argument
will be used for \code{cat(..., file = to.meme)} within \code{\link[=get_bkg]{get_bkg()}}. See
\url{http://meme-suite.org/doc/bfile-format.html} for a description of
the format.}

\item{RC}{\code{logical(1)} Calculate the background of the reverse complement
of the input sequences as well. Only valid for DNA/RNA.}

\item{list.out}{\code{logical(1)} Return background frequencies as list, with an
entry for each \code{k}. If \code{FALSE}, return a single vector.}

\item{nthreads}{\code{numeric(1)} Run \code{\link[=get_bkg]{get_bkg()}} in parallel with \code{nthreads}
threads. \code{nthreads = 0} uses all available threads.
Note that no speed up will occur for jobs with only a single sequence.}
}
\value{
If \code{to.meme = NULL} and \code{list.out = TRUE}: a list with each entry being a
named numeric vector for every element in \code{k}. If \code{to.meme = NULL} and
\code{list.out = FALSE}: a named numeric vector. Otherwise: \code{NULL}, invisibly.
}
\description{
For a set of input sequences, calculate the overall sequence background for
any k-let size. Only recommended for non-DNA/RNA sequences: otherwise use
the much faster and more efficient
\code{\link[Biostrings:nucleotideFrequency]{Biostrings::oligonucleotideFrequency()}}.
}
\examples{
## Compare to Biostrings version
library(Biostrings)
seqs.DNA <- create_sequences()
bkg.DNA <- get_bkg(seqs.DNA, k = 3, as.prob = FALSE, list.out = FALSE)
bkg.DNA2 <- oligonucleotideFrequency(seqs.DNA, 3, 1, as.prob = FALSE)
bkg.DNA2 <- colSums(bkg.DNA2)
all(bkg.DNA == bkg.DNA2)

## Create a MEME background file
get_bkg(seqs.DNA, k = 1:3, to.meme = stdout(), pseudocount = 1)

## Non-DNA/RNA/AA alphabets
seqs.QWERTY <- create_sequences("QWERTY")
bkg.QWERTY <- get_bkg(seqs.QWERTY, k = 1:2)

}
\references{
\insertRef{meme3}{universalmotif}
}
\seealso{
\code{\link[=create_sequences]{create_sequences()}}, \code{\link[=scan_sequences]{scan_sequences()}}, \code{\link[=shuffle_sequences]{shuffle_sequences()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
