% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DBscores.R
\name{make_DBscores}
\alias{make_DBscores}
\title{Create P-value databases.}
\usage{
make_DBscores(db.motifs, method = c("PCC", "MPCC", "EUCL", "MEUCL", "SW",
  "MSW", "KL", "MKL"), shuffle.db = TRUE, shuffle.k = 3,
  shuffle.method = "linear", shuffle.leftovers = NULL,
  rand.tries = 100, widths = 5:30, min.position.ic = 0,
  normalise.scores = TRUE, min.overlap = 6, min.mean.ic = 0.25,
  progress = TRUE, BP = FALSE, nthreads = 1)
}
\arguments{
\item{db.motifs}{\code{list} Database motifs.}

\item{method}{\code{character} Any of \code{c('PCC', 'MPCC', 'EUCL', 'MEUCL', 'SW', 'MSW', 'KL', 'MKL')}. See \code{\link[=compare_motifs]{compare_motifs()}}. If multiple methods
are provided, each result will be in its own list.}

\item{shuffle.db}{\code{logical(1)} Shuffle \code{db.motifs} rather than
generate random motifs with \code{\link[=create_motif]{create_motif()}}.}

\item{shuffle.k}{\code{numeric(1)} See \code{\link[=shuffle_motifs]{shuffle_motifs()}}.}

\item{shuffle.method}{\code{character(1)} See \code{\link[=shuffle_motifs]{shuffle_motifs()}}.}

\item{shuffle.leftovers}{Deprecated. Does nothing.}

\item{rand.tries}{\code{numeric(1)} Number of random motifs to create for
P-value computation.}

\item{widths}{\code{numeric} Motif widths to use in P-value database calculation.}

\item{min.position.ic}{\code{numeric(1)} Minimum information content required between
individual alignment positions for it to be counted in the final alignment
score. It is recommended to use this together with \code{normalise.scores = TRUE},
as this will help punish scores resulting from only a fraction of an
alignment.}

\item{normalise.scores}{\code{logical(1)} See \code{\link[=compare_motifs]{compare_motifs()}}.}

\item{min.overlap}{\code{numeric(1)} Minimum required motif overlap. See
\code{\link[=compare_motifs]{compare_motifs()}}.}

\item{min.mean.ic}{\code{numeric(1)} See \code{\link[=compare_motifs]{compare_motifs()}}.}

\item{progress}{\code{logical(1)} Show progress.}

\item{BP}{\code{logical(1)} Deprecated. See \code{nthreads}.}

\item{nthreads}{\code{numeric(1)} Run \code{\link[=compare_motifs]{compare_motifs()}} in parallel with \code{nthreads}
threads. \code{nthreads = 0} uses all available threads.}
}
\value{
A \code{data.frame} with score distributions for the
input database, or a \code{list} with a \code{data.frame} for each method and
an additional \code{list} entry logging function parameters if more than one
method is provided.
}
\description{
Generate data used by \code{\link[=compare_motifs]{compare_motifs()}} for P-value calculations. By default,
\code{\link[=compare_motifs]{compare_motifs()}} uses an internal database based on the JASPAR2018 core motifs
\insertCite{jaspar}{universalmotif}.
}
\examples{
\dontrun{
library(MotifDb)
motifs <- convert_motifs(MotifDb[1:100])
scores <- make_DBscores(motifs, method = "PCC")
compare_motifs(motifs, 1:100, db.scores = scores)
}

}
\references{
\insertRef{jaspar}{universalmotif}
}
\seealso{
\code{\link[=compare_motifs]{compare_motifs()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
