% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_motifs.R
\name{merge_motifs}
\alias{merge_motifs}
\title{Merge motifs using msaClustalW}
\usage{
merge_motifs(motifs, newname = "merged motif", cluster = "default",
  substitutionMatrix = "clustalw", ...)
}
\arguments{
\item{motifs}{List of motifs.}

\item{newname}{Name of final merged motif.}

\item{cluster}{nj (defaul) or upgma}

\item{substitutionMatrix}{iub or clustalw}

\item{...}{Other settings for function msaClustalW.}
}
\value{
A single motif object.
}
\description{
The 'extrainfo' slot will be dropped, as well as any non-identical slots.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
