% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_dist.R
\name{motif_dist}
\alias{motif_dist}
\title{Motif distance score matrix.}
\usage{
motif_dist(motifs, metric = "PCC", align = "SWU", gap_open = 1,
  gap_extend = 0.5)
}
\arguments{
\item{motifs}{List of motifs}

\item{metric}{Distance metric. One of PCC (Pearson Correlation Coefficient),
ALLR (Average Log-Likelihood Ratio), ALLR_LL (ALLR with low
score limit of 2), CS (Chi-Squared), KL (Kullback-Lieber),
SSD (Sum of Squared Distances).}

\item{align}{Alignment strategy. One of NW (Needleman-Wunsch), 
SW (Smith-Waterman with linear gaps), SWA (Smith-Waterman
with affine gaps), SWU (Smith-Waterman with ungapped and
alignments are extended).}

\item{gap_open}{Gap open penalty.}

\item{gap_extend}{Gap extension penaly.}
}
\description{
MotIV implementation of the STAMP motif analysis software.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
