% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_logo.R
\name{motif_logo}
\alias{motif_logo}
\title{Generate ggplot2 motif logos.}
\usage{
motif_logo(motif, engine = "ggseqlogo", type = "ic", ...)
}
\arguments{
\item{motif}{universalmotif object.}

\item{engine}{Character. ggseqlogo, seqLogo, motifStack.}

\item{type}{Character. ic or prob.}

\item{...}{Package-specific params.}
}
\description{
Create logos using ggplot2 + ggseqlogo or seqLogo.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
