% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_rc.R
\name{motif_rc}
\alias{motif_rc}
\title{Get the reverse complement of a motif.}
\usage{
motif_rc(motifs, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a single motif.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
List of motifs or single motif object.
}
\description{
Get the reverse complement of a motif.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.rc <- motif_rc(jaspar)

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
