% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_simil.R
\name{motif_simil}
\alias{motif_simil}
\title{Motif similarity matrix.}
\usage{
motif_simil(motifs, method = "Pearson")
}
\arguments{
\item{motifs}{List of motifs.}

\item{method}{One of Euclidean, Pearson, KL.}
}
\value{
Similarity matrix.
}
\description{
TFBSTools implementation of motif similarity. Supports parallel execution
via foreach package.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
