% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{motif_slots<-}
\alias{motif_slots<-}
\title{Replacement function for \linkS4class{universalmotif}.}
\usage{
motif_slots(object, slot) <- value
}
\arguments{
\item{object}{Motif object.}

\item{slot}{Character. Single slot to replace. See
\linkS4class{universalmotif} for a description of the slots.}

\item{value}{Information to replace the slot with.}
}
\value{
Motif object.
}
\description{
Replace individual slots for \linkS4class{universalmotif}.
}
\examples{
   motifs <- system.file("extdata", "minimal.meme",
                         package = "universalmotif")
   motifs <- read_motifs(motifs)
   motif <- motifs[[1]]
   ## change the name:
   motif_slots(motif, "name") <- "New_name"
   ## the 'motif' object is of 'type' PPM; changing it manually to PCM
   ## will cause an error to occur
   \dontrun{
      motif_slots(motif, "type") <- "PCM"
   }

}
\seealso{
\code{\link{convert_type}} to convert motif 'type'
}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
