% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamm.R
\name{read_bamm}
\alias{read_bamm}
\title{Import BaMM motifs.}
\usage{
read_bamm(file, alphabet = "DNA", name = "motif", skip = 0)
}
\arguments{
\item{file}{Character.}

\item{alphabet}{Character.}

\item{name}{Character.}

\item{skip}{Numeric.}
}
\value{
A universalmotif object.
}
\description{
Import BaMM motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
