% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cisbp.R
\name{read_cisbp}
\alias{read_cisbp}
\title{Import CIS-BP motifs.}
\usage{
read_cisbp(file, skip = 0, BPPARAM = SerialParam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
List of universalmotif objects.
}
\description{
Import CIS-BP motifs.
}
\examples{
cisbp <- read_cisbp(system.file("extdata", "cisbp.txt",
                                package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
