% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_motifs.R
\name{read_motifs}
\alias{read_motifs}
\title{Import universalmotif formatted motifs.}
\usage{
read_motifs(file, skip = 0)
}
\arguments{
\item{file}{\code{character(1)} File name.}

\item{skip}{\code{numeric(1)} If not zero, will skip however many desired lines in the
file before starting to read.}
}
\value{
\code{list} \linkS4class{universalmotif} objects.
}
\description{
Import motifs created from \code{\link[=write_motifs]{write_motifs()}}. For optimal storage of
\code{universalmotif} class motifs, consider using \code{\link[=saveRDS]{saveRDS()}} and
\code{\link[=readRDS]{readRDS()}}. The \code{universalmotif} format will not be documented,
as realistically the need to generate these manually/elsewhere should
be nonexistent.
}
\seealso{
Other read_motifs: \code{\link{read_cisbp}},
  \code{\link{read_homer}}, \code{\link{read_jaspar}},
  \code{\link{read_matrix}}, \code{\link{read_meme}},
  \code{\link{read_transfac}}, \code{\link{read_uniprobe}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{read_motifs}
