% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uniprobe.R
\name{read_uniprobe}
\alias{read_uniprobe}
\title{Import UNIPROBE motifs.}
\usage{
read_uniprobe(file, skip = 0, BPPARAM = SerialParam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric. If not zero, will skip however many desired lines in the
file before starting to read.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
List of universalmotif objects.
}
\description{
Import UNIPROBE formatted motifs. Assumed DNA.
}
\examples{
uniprobe.minimal <- read_uniprobe(system.file("extdata", "uniprobe_minimal.txt",
                                              package = "universalmotif"))
uniprobe.full <- read_uniprobe(system.file("extdata", "uniprobe_full.txt",
                                           package = "universalmotif"))

}
\references{
\insertRef{uniprobe}{universalmotif}
}
\seealso{
Other read_motifs: \code{\link{read_cisbp}},
  \code{\link{read_homer}}, \code{\link{read_jaspar}},
  \code{\link{read_matrix}}, \code{\link{read_meme}},
  \code{\link{read_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
