% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sites.R
\name{sample_sites}
\alias{sample_sites}
\title{Generate binding sites from a motif.}
\usage{
sample_sites(motif, n = 100, BPPARAM = SerialParam())
}
\arguments{
\item{motif}{Motif object.}

\item{n}{Numeric. Number of sites to generate.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
\linkS4class{XStringSet} object.
}
\description{
Generate binding sites from a motif.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
