% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_sequences.R
\name{shuffle_sequences}
\alias{shuffle_sequences}
\title{Shuffle input sequences or generate random sequences.}
\usage{
shuffle_sequences(sequences, inter = FALSE, alphabet, bkg, numseqs = 100,
  seqlen = 100)
}
\arguments{
\item{sequences}{XStringSet objects.}

\item{inter}{Logical. Shuffle letters between sequences? If \code{FALSE},
then the alphabet frequencies for each sequence are maintained.
Ignored if \code{sequences} is missing.}

\item{alphabet}{Character. Ignored if \code{sequences} is given. If not one
of 'DNA', 'RNA', or 'AA', then a sequence of letters
representing the alphabet to use (e.g. "QWERTY").}

\item{bkg}{Numeric. Ignored if \code{sequences} is given; otherwise generate
random sequences with these alphabet frequencies. If missing,
will generate sequences with even background frequencies.}

\item{numseqs}{Numeric. Ignored if \code{sequences} is given; otherwise
the number of random sequences to generate.}

\item{seqlen}{Numeric. Ignored if \code{sequences} is given;otherwise the
length of the random sequences.}
}
\value{
XStringSet object.
}
\description{
Shuffle input sequences or generate random sequences.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
