% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_alph.R
\name{switch_alph}
\alias{switch_alph}
\title{Switch between DNA and RNA alphabets.}
\usage{
switch_alph(motifs, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{Motif object or list of.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
The DNA/RNA version of the motif(s).
}
\description{
Switch between DNA and RNA alphabets.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
