% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif_df.R
\name{tidy-motifs}
\alias{tidy-motifs}
\alias{to_df}
\alias{update_motifs}
\alias{to_list}
\title{Tidy manipulation of motifs.}
\usage{
to_df(motifs)

update_motifs(motif_df)

to_list(motif_df)
}
\arguments{
\item{motifs}{List of motifs.}

\item{motif_df}{Motif \code{data.frame} generated by \code{\link[=to_df]{to_df()}}.}
}
\value{
For \code{\link[=to_df]{to_df()}}: a \code{data.frame} with the exposed slots as columns.

For \code{\link[=update_motifs]{update_motifs()}}: the updated \code{data.frame}.

For \code{\link[=to_list]{to_list()}}: a list of motifs.
}
\description{
Tidy manipulation of motifs.
}
\examples{
\dontrun{
library(universalmotif)
library(dplyr)

m <- c(create_motif(name = "motif A"), create_motif(name = "motif B"))
# Change the names of the motifs using the tidy way:
m <- m \%>\%
   to_df() \%>\%
   mutate(name = paste0(name, "-2")) \%>\%
   to_list()
}

}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
