% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_motif.R
\name{trim_motif}
\alias{trim_motif}
\title{Trim motifs.}
\usage{
trim_motif(motif, IC_cutoff = 0.25)
}
\arguments{
\item{motif}{Motif object.}

\item{IC_cutoff}{Numeric. Minimum allowed information content.}
}
\value{
Motif object.
}
\description{
Trim motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
