% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_motifs.R
\name{trim_motifs}
\alias{trim_motifs}
\title{Trim motifs.}
\usage{
trim_motifs(motifs, IC_cutoff = 0.25)
}
\arguments{
\item{motifs}{Motif object.}

\item{IC_cutoff}{Numeric. Minimum allowed information content.}
}
\value{
Motifs Motif object or list.
}
\description{
Trim motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
