% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umot-class.R
\docType{class}
\name{umot-class}
\alias{umot-class}
\title{[UNDER CONSTRUCTION] umot: Motif class.}
\description{
Placeholder description.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. Motif name.}

\item{\code{motif}}{Matrix. Contains motif.}

\item{\code{alphabet}}{Character. Describes alphabet.}

\item{\code{letters}}{Character. Alphabet letters.}

\item{\code{type}}{Character. Describes how the motif is stores; i.e., as a PWM,
etc.}

\item{\code{nsites}}{Integer. Number of times motif was found in original set.}

\item{\code{pval}}{Double. Motif p-value, if present.}

\item{\code{pseudoweights}}{Numeric vector. Motif pseudoweights.}

\item{\code{eval}}{Double. Motif e-value, if present.}
}}

