% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R,
%   R/universalmotif-methods.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\alias{universalmotif}
\alias{[,universalmotif,ANY,ANY,ANY-method}
\alias{[<-,universalmotif,ANY,ANY,ANY-method}
\alias{initialize,universalmotif-method}
\alias{show,universalmotif-method}
\title{universalmotif: Motif class.}
\usage{
\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i, drop = "missing")

\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i) <- value

\S4method{initialize}{universalmotif}(.Object, name, altname, family, organism,
  motif, alphabet = "DNA", type, icscore, nsites, pseudoweight = 0.8, bkg,
  bkgsites, consensus, strand = "+-", pval, qval, eval, extrainfo)

\S4method{show}{universalmotif}(object)
}
\value{
A motif object.
}
\description{
Description.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Accessor function.

\item \code{[<-}: Replacement method.

\item \code{initialize}: Initialize method.

\item \code{show}: Show method.
}}

\section{Slots}{

\describe{
\item{\code{name}}{Character. Motif name.}

\item{\code{altname}}{Character.}

\item{\code{family}}{Character.}

\item{\code{organism}}{Character.}

\item{\code{motif}}{Matrix. Contains motif.}

\item{\code{alphabet}}{Character. Can be: DNA, RNA, AA, custom.}

\item{\code{type}}{Character. Can be: PCM, PPM, PWM.}

\item{\code{icscore}}{Numeric. Total information content for all positions.}

\item{\code{nsites}}{Numeric. Total number of sites.}

\item{\code{pseudoweight}}{Numeric. Amount of smoothing to apply.}

\item{\code{bkg}}{Numeric. Background letter frequencies.}

\item{\code{bkgsites}}{Numeric.}

\item{\code{consensus}}{Character. Motif consensus sequence.}

\item{\code{strand}}{Character. '+' and/or '-'.}

\item{\code{pval}}{Numeric.}

\item{\code{qval}}{Numeric.}

\item{\code{eval}}{Numeric.}

\item{\code{extrainfo}}{List.}
}}

\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
