% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R, R/utils.R
\name{make_DBscores}
\alias{make_DBscores}
\alias{utilities}
\alias{ppm_to_icm}
\alias{icm_to_ppm}
\alias{pcm_to_ppm}
\alias{ppm_to_pcm}
\alias{ppm_to_pwm}
\alias{pwm_to_ppm}
\alias{position_icscore}
\alias{get_consensus}
\alias{consensus_to_ppm}
\alias{consensus_to_ppmAA}
\alias{get_consensusAA}
\alias{summarise_motifs}
\title{Utility functions.}
\usage{
make_DBscores(db.motifs, method, shuffle.db = TRUE, shuffle.k = 3,
  shuffle.method = "linear", shuffle.leftovers = "asis",
  rand.tries = 1000, normalise.scores = TRUE, min.overlap = 6,
  min.mean.ic = 0, progress = TRUE, BP = FALSE)

ppm_to_icm(position, bkg, schneider_correction = FALSE, nsites,
  relative_entropy = FALSE)

icm_to_ppm(position)

pcm_to_ppm(position, pseudocount = 0.8)

ppm_to_pcm(position, nsites = 100)

ppm_to_pwm(position, bkg, pseudocount = 0.8, nsites = 100,
  smooth = TRUE)

pwm_to_ppm(position, bkg)

position_icscore(position, bkg, type, pseudocount = 0.8, nsites = 100,
  relative_entropy = FALSE)

get_consensus(position, alphabet = "DNA", type = "PPM",
  pseudocount = 0.8)

consensus_to_ppm(letter)

consensus_to_ppmAA(letter)

get_consensusAA(position, type, pseudocount)

summarise_motifs(motifs, na.rm = TRUE)
}
\arguments{
\item{db.motifs}{\code{list} Database motifs.}

\item{method}{\code{character(1)} One of \code{c('Pearson', 'Euclidean', 'KL')}.}

\item{shuffle.db}{\code{logical(1)} Shuffle \code{db.motifs} rather than
generate random motifs with \code{\link[=create_motif]{create_motif()}}.}

\item{shuffle.k}{\code{numeric(1)}}

\item{shuffle.method}{\code{character(1)}}

\item{shuffle.leftovers}{\code{character(1)}}

\item{rand.tries}{\code{numeric(1)}}

\item{normalise.scores}{\code{logical(1)}}

\item{min.overlap}{\code{numeric(1)} Minimum required motif overlap.}

\item{min.mean.ic}{\code{numeric(1)}}

\item{progress}{\code{logical(1)} Show progress.}

\item{BP}{\code{logical(1)} Use BiocParallel.}

\item{position}{\code{numeric} A numeric vector representing the frequency or
probability for each alphabet letter at a specific position.}

\item{bkg}{\code{Numeric} Should be the same length as the alphabet length.}

\item{schneider_correction}{\code{logical(1)} Apply sample size correction.}

\item{nsites}{\code{numeric(1)} Number of sites motif originated from.}

\item{relative_entropy}{\code{logical(1)} Calculate information content as
relative entropy or Kullback-Leibler divergence.}

\item{pseudocount}{\code{numeric(1)} Used to prevent zeroes in motif matrix.}

\item{smooth}{\code{logical(1)} Apply pseudocount correction.}

\item{type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM' 'ICM')}.}

\item{alphabet}{\code{character(1)} One of \code{c('DNA', 'RNA')}.}

\item{letter}{\code{character(1)} Any DNA, RNA, or AA IUPAC letter. Ambiguity letters
are accepted.}

\item{motifs}{\code{list} A list of \linkS4class{universalmotif} motifs.}

\item{na.rm}{\code{logical} Remove columns where all values are \code{NA}.}
}
\value{
For \code{\link[=ppm_to_icm]{ppm_to_icm()}}, \code{\link[=icm_to_ppm]{icm_to_ppm()}}, \code{\link[=pcm_to_ppm]{pcm_to_ppm()}},
\code{\link[=ppm_to_pcm]{ppm_to_pcm()}}, \code{\link[=ppm_to_pwm]{ppm_to_pwm()}}, and \code{\link[=pwm_to_ppm]{pwm_to_ppm()}}: a \code{numeric}
vector with length equal to input \code{numeric} vector.

For \code{\link[=consensus_to_ppm]{consensus_to_ppm()}} and \code{\link[=consensus_to_ppmAA]{consensus_to_ppmAA()}}: a numeric
vector of length 4 and 20, respectively.

For \code{\link[=position_icscore]{position_icscore()}}: a \code{numeric} vector of length 1.

For \code{\link[=get_consensus]{get_consensus()}} and \code{\link[=get_consensusAA]{get_consensusAA()}}: a character vector
of length 1.

For \code{\link[=make_DBscores]{make_DBscores()}}: a \code{data.frame} with score distributions for the
input database.

For \code{\link[=summarise_motifs]{summarise_motifs()}}: a \code{data.frame} with columns representing
the \linkS4class{universalmotif} slots.
}
\description{
Various small functions used for motif creation.
}
\seealso{
\code{\link[=create_motif]{create_motif()}}, \code{\link[=compare_motifs]{compare_motifs()}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
