% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transfac.R
\name{write_transfac}
\alias{write_transfac}
\title{Export motifs in TRANSFAC format.}
\usage{
write_transfac(motifs, file, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs in TRANSFAC format.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                    package = "universalmotif"))
write_transfac(jaspar, tempfile())

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
