% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_multifreq.R
\name{add_multifreq}
\alias{add_multifreq}
\title{Add multi-nucleotide information to a motif.}
\usage{
add_multifreq(motif, sequences, add.k = 2:3, RC = FALSE, threshold = 0.01,
  threshold.type = "logodds", motifs.perseq = 1, BPPARAM = SerialParam())
}
\arguments{
\item{motif}{universalmotif object.}

\item{sequences}{DNAStringSet.}

\item{add.k}{Numeric.}

\item{RC}{Logical.}

\item{threshold}{Numeric.}

\item{threshold.type}{Character.}

\item{motifs.perseq}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\description{
Add multi-nucleotide information to a motif.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
