% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArabidopsisMotif}
\alias{ArabidopsisMotif}
\title{Arabidopsis motif.}
\format{\linkS4class{universalmotif}}
\usage{
ArabidopsisMotif
}
\description{
Arabidopsis motif trained from \code{\link{ArabidopsisPromoters}} using
MEME v4.
}
\keyword{datasets}
