% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R
\name{compare_motifs}
\alias{compare_motifs}
\title{Compare motifs.}
\usage{
compare_motifs(motifs, method = "Euclidean", BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs.}

\item{method}{One of Euclidean, Pearson, and KL.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Distance matrix for Euclidean and KL; similarity matrix for Pearson.
}
\description{
TFBSTools implementation of motif comparison. Rewritten in C++.
}
\references{
\insertRef{tfbstools}{universalmotif}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
