% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R
\name{compare_motifs}
\alias{compare_motifs}
\title{Compare motifs.}
\usage{
compare_motifs(motifs, compare.to, db.scores, k = 1, use.type = "PPM",
  method = "Pearson", tryRC = TRUE, min.overlap = 6, min.mean.ic = 0.5,
  relative_entropy = FALSE, max.p = 0.05, max.e = 10,
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs. If not a \linkS4class{universalmotif} object,
they will be converted. See \code{\link{convert_motifs}} for supported
classes.}

\item{compare.to}{Numeric. If NULL, compares all motifs to all other motifs.
Otherwise compares all motifs to the specified motif(s).}

\item{db.scores}{data.frame.}

\item{k}{Numeric.}

\item{use.type}{Character. One of 'PCM' (Pearson only), 'PPM' (any method),
'PWM' (Pearson only), and 'ICM' (any method). The latter
two allow for taking into account the background frequencies 
(for ICM, only if \code{relative_entropy = TRUE}).}

\item{method}{One of 'Euclidean', 'Pearson', and 'KL'.}

\item{tryRC}{Try the reverse complement of the motifs as well, report the
best score.}

\item{min.overlap}{Numeric. Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{Numeric.}

\item{relative_entropy}{Logical.}

\item{max.p}{Numeric.}

\item{max.e}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Distance matrix for Euclidean or KL; similarity matrix for Pearson.
}
\description{
TFBSTools implementation of motif comparison, rewritten in C++. See
\code{\link[TFBSTools]{PWMSimilarity}} for details.
}
\details{
The implementations of this function are the exact same as that of the
   \code{\link[TFBSTools]{PWMSimilarity}} function, except that this function
   allows for more than two motifs to be compared as well as providing
   significant performance gains.
}
\examples{
motif1 <- create_motif()
motif2 <- create_motif()
motif1vs2 <- compare_motifs(list(motif1, motif2))
# to get a dist object:
as.dist(motif1vs2)

}
\references{
\insertRef{tfbstools}{universalmotif}
}
\seealso{
\code{\link{convert_motifs}}, \code{\link[TFBSTools]{PWMSimilarity}},
   \code{\link{motif_tree}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
