% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R
\name{compare_motifs}
\alias{compare_motifs}
\title{Compare motifs.}
\usage{
compare_motifs(motifs, compare.to, db.scores, use.freq = 1,
  use.type = "PPM", method = "MPCC", tryRC = TRUE, min.overlap = 6,
  min.mean.ic = 0.5, relative_entropy = FALSE,
  normalise.scores = FALSE, max.p = 0.01, max.e = 10,
  progress_bar = FALSE)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable motif formats.}

\item{compare.to}{\code{numeric} If missing, compares all motifs to all other motifs.
Otherwise compares all motifs to the specified motif(s).}

\item{db.scores}{\code{data.frame} See \code{details}.}

\item{use.freq}{\code{numeric(1)}. For comparing the \code{multifreq} slot.}

\item{use.type}{\code{character(1)} One of code{'PPM'} and \code{'ICM'}.
The latter allows for taking into account the background
frequencies if \code{relative_entropy = TRUE}.}

\item{method}{\code{character(1)} One of \code{c('PCC', 'MPCC', 'EUCL', 'MEUCL',
'SW', 'MSW', 'KL', 'MKL')}. See details.}

\item{tryRC}{\code{logical} Try the reverse complement of the motifs as well,
report the best score.}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{\code{numeric(1)} Minimum information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{relative_entropy}{\code{logical(1)} For ICM calculation. See
\code{\link{convert_type}}.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions, as well as alignments between motifs of similar length.
Similarity scores are multiplied by the ratio of
aligned positions to the total number of positions in the larger motif,
and the inverse for distance scores.}

\item{max.p}{\code{numeric(1)} Maximum P-value allowed in reporting matches.
Only used if \code{compare.to} is set.}

\item{max.e}{\code{numeric(1)} Maximum E-value allowed in reporting matches.
Only used if \code{compare.to} is set. The E-value is the P-value multiplied
by the number of input motifs times two.}

\item{progress_bar}{\code{logical(1)} Show progress.}
}
\value{
\code{matrix} if \code{compare.to} is missing; \code{data.frame} otherwise.
 \itemize{
   \item PCC: 0 represents complete distance, >0 similarity.
   \item MPCC: 0 represents complete distance, 1 complete similarity.
   \item EUCL: 0 represents complete similarity, >0 distance.
   \item MEUCL: 0 represents complete similarity, sqrt(2) complete distance.
   \item SW: 0 represents complete distance, >0 similarity.
   \item MSW: 0 represents complete distance, 2 complete similarity.
   \item KL: 0 represents complete similarity, >0 distance.
   \item MKL: 0 represents complete similarity, 4.62 complete distance.
 }
}
\description{
Compare motifs using four metrics: Pearson correlation coefficient,
Euclidean distance, Sandelin-Wasserman similarity, and Kullback-Leibler
divergence.
}
\details{
Comparisons are calculated between two motifs at a time. All possible alignments
are scored, and the best score is reported. Scores are calculated per position
and summed, unless the 'mean' version of the specific metric is chosen. If using
a similarity metric, then the sum of scores will favour comparisons between
longer motifs; and for distance metrics, the sum of scores will favour
comparisons between short motifs. This can be avoided by using the 'mean' of
scores.

\itemize{
  \item PCC: Pearson correlation coefficient
  \item MPCC: Mean PCC
  \item EUCL: Euclidian distance
  \item MEUCL: Mean EUCL
  \item SW: Sandelin-Wasserman similarity
  \item MSW: Mean SW
  \item KL: Kullback-Leibler divergence
  \item MKL: Mean Kullback-Leibler divergence
}

To note regarding p-values: p-values are pre-computed using the
\code{make_DBscores} function. If not given, then uses a set of internal
precomputed p-values from the JASPAR2018 CORE motifs. These precalculated
scores are dependent on the length of the motifs being compared; this takes
into account that comparing small motifs with larger motifs leads to higher
scores, since the probability of finding a higher scoring alignment is
higher.

The default p-values have been precalculated for regular DNA motifs; they
are of little use for motifs with a different number of alphabet letters
(or even the \code{multifreq} slot).
}
\examples{
motif1 <- create_motif()
motif2 <- create_motif()
motif1vs2 <- compare_motifs(list(motif1, motif2), method = "MPCC")
## to get a dist object:
as.dist(1 - motif1vs2)

}
\references{
\insertRef{jaspar}{universalmotif}

   \insertRef{tfbstools}{universalmotif}
}
\seealso{
\code{\link{convert_motifs}}, \code{\link[TFBSTools]{PWMSimilarity}},
   \code{\link{motif_tree}}, \code{\link{view_motifs}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
