% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R
\name{compare_motifs}
\alias{compare_motifs}
\title{Compare motifs.}
\usage{
compare_motifs(motifs, compare.to, db.scores, use.freq = 1,
  use.type = "PPM", method = "MPCC", tryRC = TRUE, min.overlap = 6,
  min.mean.ic = 0.25, min.position.ic = 0, relative_entropy = FALSE,
  normalise.scores = FALSE, max.p = 0.01, max.e = 10,
  progress = FALSE, BP = FALSE, nthreads = 1, db.version = 2)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{compare.to}{\code{numeric} If missing, compares all motifs to all other motifs.
Otherwise compares all motifs to the specified motif(s).}

\item{db.scores}{\code{data.frame} See \code{details}.}

\item{use.freq}{\code{numeric(1)}. For comparing the \code{multifreq} slot.}

\item{use.type}{\code{character(1)} One of \code{'PPM'} and \code{'ICM'}.
The latter allows for taking into account the background
frequencies if \code{relative_entropy = TRUE}.}

\item{method}{\code{character(1)} One of \code{c('PCC', 'MPCC', 'EUCL', 'MEUCL', 'SW', 'MSW', 'KL', 'MKL', 'ALLR', 'MALLR')}. See details.}

\item{tryRC}{\code{logical(1)} Try the reverse complement of the motifs as well,
report the best score.}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{\code{numeric(1)} Minimum mean information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{min.position.ic}{\code{numeric(1)} Minimum information content required between
individual alignment positions for it to be counted in the final alignment
score. It is recommended to use this together with \code{normalise.scores = TRUE},
as this will help punish scores resulting from only a fraction of an
alignment.}

\item{relative_entropy}{\code{logical(1)} Change the ICM calculation affecting
\code{min.position.ic} and \code{min.mean.ic}. See \code{\link[=convert_type]{convert_type()}}.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions, as well as alignments between motifs of similar length.
Similarity scores are multiplied by the ratio of
aligned positions to the total number of positions in the larger motif,
and the inverse for distance scores.}

\item{max.p}{\code{numeric(1)} Maximum P-value allowed in reporting matches.
Only used if \code{compare.to} is set.}

\item{max.e}{\code{numeric(1)} Maximum E-value allowed in reporting matches.
Only used if \code{compare.to} is set. The E-value is the P-value multiplied
by the number of input motifs times two.}

\item{progress}{\code{logical(1)} Deprecated. Does nothing.}

\item{BP}{\code{logical(1)} Deprecated. See \code{nthreads}.}

\item{nthreads}{\code{numeric(1)} Run \code{\link[=compare_motifs]{compare_motifs()}} in parallel with \code{nthreads}
threads. \code{nthreads = 0} uses all available threads.}

\item{db.version}{\code{numeric(1)} Select internal P-value database. \code{1} is the
database created in v1.0.0 of the package, and \code{2} is the database
created in v1.4.0 of the package.}
}
\value{
\code{matrix} if \code{compare.to} is missing; \code{data.frame} otherwise.
\itemize{
\item PCC: The number of positions in the shorter motif represents the max possible
score. Values below that represent dissimilarity.
\item MPCC: 1 represents complete similarity, <1 dissimilarity.
\item EUCL: 0 represents complete similarity, >0 distance.
\item MEUCL: 0 represents complete similarity, sqrt(2) complete distance.
\item SW: The number of positions in the shorter motif times two represents the
max possible score. Values below that represent dissimilarity.
\item MSW: 0 represents complete distance, 2 complete similarity.
\item KL: 0 represents complete similarity, >0 distance.
\item MKL: 0 represents complete similarity, >0 complete distance.
}
}
\description{
Compare motifs using four available metrics: Pearson correlation coefficient
\insertCite{pearson}{universalmotif}, Euclidean distance
\insertCite{euclidean}{universalmotif}, Sandelin-Wasserman similarity
\insertCite{wasserman}{universalmotif}, and Kullback-Leibler divergence
\insertCite{kldiv}{universalmotif}.
}
\details{
Comparisons are calculated between two motifs at a time. All possible alignments
are scored, and the best score is reported. Scores are calculated per position
and summed, unless the 'mean' version of the specific metric is chosen. If using
a similarity metric, then the sum of scores will favour comparisons between
longer motifs; and for distance metrics, the sum of scores will favour
comparisons between short motifs. This can be avoided by using the 'mean' of
scores.
\itemize{
\item PCC: Pearson correlation coefficient

Per position:

\code{num = nrow * sum(pos1 * pos2) - sum(pos1) * sum(pos2)}

\code{denom = sqrt((nrow * sum(pos1)^2 - sum(pos1^2)) * (nrow * sum(pos2)^2 - sum(pos2^2)))}

\code{PCC = num / denom}
\item MPCC: Mean PCC

\code{MPCC = mean(PCC)}
\item EUCL: Euclidian distance

Per position:

\code{EUCL = sqrt(sum((col1 - col2)^2)) / sqrt(2)}
\item MEUCL: Mean EUCL

\code{MEUCL = sum(EUCL) / ncol(alignment)}
\item SW: Sandelin-Wasserman similarity

Per position:

\code{SW = 2 - sum((col1 - col2)^2)}
\item MSW: Mean SW

\code{MSW = mean(SW)}
\item KL: Kullback-Leibler divergence

Per position:

\code{KL = 0.5 * sum(col1 * log(col1/col2) + col2 * log(col2/col1))}
\item MKL: Mean Kullback-Leibler divergence

\code{MKL = mean(KL)}
}

To note regarding p-values: p-values are pre-computed using the
\code{make_DBscores} function. If not given, then uses a set of internal
precomputed p-values from the JASPAR2018 CORE motifs. These precalculated
scores are dependent on the length of the motifs being compared; this takes
into account that comparing small motifs with larger motifs leads to higher
scores, since the probability of finding a higher scoring alignment is
higher.

The default p-values have been precalculated for regular DNA motifs; they
are of little use for motifs with a different number of alphabet letters
(or even the \code{multifreq} slot).
}
\examples{
motif1 <- create_motif()
motif2 <- create_motif()
motif1vs2 <- compare_motifs(list(motif1, motif2), method = "MPCC")
## to get a dist object:
as.dist(1 - motif1vs2)

}
\references{
\insertRef{euclidean}{universalmotif}

\insertRef{jaspar}{universalmotif}

\insertRef{pearson}{universalmotif}

\insertRef{kldiv}{universalmotif}

\insertRef{wasserman}{universalmotif}
}
\seealso{
\code{\link[=convert_motifs]{convert_motifs()}}, \code{\link[=motif_tree]{motif_tree()}}, \code{\link[=view_motifs]{view_motifs()}},
\code{\link[=make_DBscores]{make_DBscores()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
