% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{convert_motifs}
\alias{convert_motifs}
\alias{convert_motifs,list-method}
\alias{convert_motifs,universalmotif-method}
\alias{convert_motifs,ANY-method}
\title{Convert motif class.}
\usage{
convert_motifs(motifs, class = "universalmotif-universalmotif",
  BPPARAM = bpparam())

\S4method{convert_motifs}{list}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())

\S4method{convert_motifs}{universalmotif}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())

\S4method{convert_motifs}{ANY}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())
}
\arguments{
\item{motifs}{Single motif object or list.}

\item{class}{Desired motif class. E.g. 'motifStack-pfm'.}

\item{BPPARAM}{Param for bplapply.}
}
\value{
Single motif object or list.
}
\description{
Convert motif class.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Convert a list of motifs.

\item \code{universalmotif}: Convert a universalmotif object.

\item \code{ANY}: Convert non-universalmotif class motifs.
}}

\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.motifstack.pfm <- convert_motifs(jaspar, "motifStack-pfm")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
