% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/convert_motifs-methods.R,
%   R/methods.R
\docType{methods}
\name{convert_motifs}
\alias{convert_motifs}
\alias{convert_motifs,universalmotif-method}
\alias{convert_motifs,list-method}
\title{[UNDER CONSTRUCTION] Convert motifs between various classes.}
\usage{
convert_motifs(motif, out_class, ...)

\S4method{convert_motifs}{universalmotif}(motif, out_class,
  pseudoweight = 0.8, background = c(A = 0.25, C = 0.25, G = 0.25, T =
  0.25))

\S4method{convert_motifs}{list}(motif, out_class, ...)
}
\arguments{
\item{motif}{Motif object or list of motif objects.}

\item{out_class}{Character. Desired motif class to convert to.}
}
\value{
Motif object in the desired class, or list of motifs.
}
\description{
Convert motif objects between various supported classes; some of these
require their respective packages being installed.
}
\section{Methods (by class)}{
\itemize{
\item \code{universalmotif}: Convert from universalmotif.

\item \code{list}: Convert a list of motifs.
}}

\examples{
  motifs <- system.file("extdata", "minimal.meme", package = "universalmotif")
  motifs <- read_meme(motifs, out_class = "matrix-2")
  motifs <- convert_motifs(motifs, out_class = "matrix-1")

}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
