% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_motifs.R
\name{convert_motifs}
\alias{convert_motifs}
\title{[UNDER CONSTRUCTION] Convert motifs between various classes.}
\usage{
convert_motifs(motif_list, out_class = "matrix=1")
}
\arguments{
\item{motif_list}{List of motif objects.}

\item{out_class}{Character. Desired motif class to convert to.}
}
\value{
List of motif objects in the desired class.
}
\description{
Convert motif objects between various supported classes; some of these
require their respective packages being installed.
}
\examples{
  motifs <- system.file("extdata", "minimal.meme", package = "universalmotif")
  motifs <- read_meme(motifs, show_warnings = FALSE, out_class = "matrix-2")
  motifs <- convert_motifs(motifs, out_class = "matrix-1")

}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
