% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_motifs.R
\docType{methods}
\name{convert_motifs}
\alias{convert_motifs}
\alias{convert_motifs,list-method}
\alias{convert_motifs,universalmotif-method}
\alias{convert_motifs,MotifList-method}
\alias{convert_motifs,TFFMFirst-method}
\alias{convert_motifs,PFMatrix-method}
\alias{convert_motifs,PWMatrix-method}
\alias{convert_motifs,ICMatrix-method}
\alias{convert_motifs,XMatrixList-method}
\alias{convert_motifs,pwm-method}
\alias{convert_motifs,pcm-method}
\alias{convert_motifs,pfm-method}
\alias{convert_motifs,PWM-method}
\alias{convert_motifs,Motif-method}
\alias{convert_motifs,matrix-method}
\title{Convert motif class.}
\usage{
convert_motifs(motifs, class = "universalmotif-universalmotif")

\S4method{convert_motifs}{list}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{universalmotif}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{MotifList}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{TFFMFirst}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{PFMatrix}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{PWMatrix}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{ICMatrix}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{XMatrixList}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{pwm}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{pcm}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{pfm}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{PWM}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{Motif}(motifs,
  class = "universalmotif-universalmotif")

\S4method{convert_motifs}{matrix}(motifs,
  class = "universalmotif-universalmotif")
}
\arguments{
\item{motifs}{Single motif object or list.}

\item{class}{\code{character(1)} Desired motif class. Input as
'package-class'. If left empty, defaults to
'universalmotif-universalmotif'. (See details.)}
}
\value{
Single motif object or list.
}
\description{
Allows for easy transfer of motif information between different classes
defined by other Bioconductor packages. This function is also used by
nearly all other functions in this package; so any motifs of a compatible
class can be used without needed to convert beforehand.
}
\details{
The following packge-class combinations can be used as input:
   \itemize{
      \item MotifDb-MotifList
      \item TFBSTools-PFMatrix
      \item TFBSTools-PWMatrix
      \item TFBSTools-ICMatrix
      \item TFBSTools-PFMatrixList
      \item TFBSTools-PWMatrixList
      \item TFBSTools-ICMatrixList
      \item TFBSTools-TFFMFirst
      \item seqLogo-pwm
      \item motifStack-pcm
      \item motifStack-pfm
      \item PWMEnrich-PWM
      \item motifRG-Motif
      \item universalmotif-universalmotif
      \item matrix
   }

   The following package-class combinations can be output:
   \itemize{
      \item MotIV-pwm2
      \item TFBSTools-PFMatrix
      \item TFBSTools-PWMatrix
      \item TFBSTools-ICMatrix
      \item TFBSTools-TFFMFirst
      \item seqLogo-pwm
      \item motifStack-pcm
      \item motifStack-pfm
      \item PWMEnrich-PWM
      \item Biostrings-PWM (\code{type = 'log2prob'})
      \item rGADEM-motif
      \item universalmotif-universalmotif
   }
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Convert a list of motifs.

\item \code{universalmotif}: Convert a \linkS4class{universalmotif} object.

\item \code{MotifList}: Convert MotifList motifs.

\item \code{TFFMFirst}: Convert TFFMFirst motifs.

\item \code{PFMatrix}: Convert PFMatrix motifs.

\item \code{PWMatrix}: Convert PWMatrix motifs.

\item \code{ICMatrix}: Convert ICMatrix motifs.

\item \code{XMatrixList}: Convert XMatrixList motifs.

\item \code{pwm}: Convert pwm motifs.

\item \code{pcm}: Convert pcm motifs.

\item \code{pfm}: Convert pfm motifs.

\item \code{PWM}: Convert PWM motifs.

\item \code{Motif}: Convert Motif motifs.

\item \code{matrix}: Create motif from matrices.
}}

\examples{
# convert from universalmotif:
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
if (requireNamespace("motifStack", quietly = TRUE)) {
  jaspar.motifstack.pfm <- convert_motifs(jaspar, "motifStack-pfm")
}

# convert from another class to universalmotif:
library(TFBSTools)
data(MA0003.2)
motif <- convert_motifs(MA0003.2)

# convert from another class to another class
if (requireNamespace("PWMEnrich", quietly = TRUE)) {
  motif <- convert_motifs(MA0003.2, "PWMEnrich-PWM")
}

# the 'convert_motifs' function is embedded in the rest of the universalmotif
# functions; non-universalmotif class motifs can be used
MA0003.2.trimmed <- trim_motifs(MA0003.2)
# note: if the motif object going in has information that the
# 'universalmotif' class can't hold, it will be lost

}
\references{
\insertRef{seqlogo}{universalmotif}

   \insertRef{rgadem}{universalmotif}

   \insertRef{motiv}{universalmotif}

   \insertRef{motifstack}{universalmotif}

   \insertRef{biostrings}{universalmotif}

   \insertRef{motifdb}{universalmotif}

   \insertRef{pwmenrich}{universalmotif}

   \insertRef{tfbstools}{universalmotif}

   \insertRef{motifrg}{universalmotif}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
