% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{convert_motifs}
\alias{convert_motifs}
\alias{convert_motifs,list-method}
\alias{convert_motifs,universalmotif-method}
\alias{convert_motifs,ANY-method}
\title{Convert motif class.}
\usage{
convert_motifs(motifs, class = "universalmotif-universalmotif",
  BPPARAM = bpparam())

\S4method{convert_motifs}{list}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())

\S4method{convert_motifs}{universalmotif}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())

\S4method{convert_motifs}{ANY}(motifs,
  class = "universalmotif-universalmotif", BPPARAM = bpparam())
}
\arguments{
\item{motifs}{Single motif object or list.}

\item{class}{Desired motif class. Input as 'package-class'. If left empty,
defaults to 'universalmotif-universalmotif'. (See details.)}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Single motif object or list.
}
\description{
Allows for easy transfer of motif information between different classes
defined by other Bioconductor packages. This function is also used by
most other functions, which allows them to be used with foreign class
types seamlessly.
}
\details{
The following packge-class combinations can be used as input:
   \itemize{
      \item MotifDb-MotifList
      \item TFBSTools-PFMatrix
      \item TFBSTools-PWMatrix
      \item TFBSTools-ICMatrix
      \item TFBSTools-PFMatrixList
      \item TFBSTools-PWMatrixList
      \item TFBSTools-ICMatrixList
      \item seqLogo-pwm
      \item motifStack-pcm
      \item motifStack-pfm
      \item PWMEnrich-PWM
      \item motifRG-Motif
      \item universalmotif-universalmotif
   }

   The following package-class combinations can be output:
   \itemize{
      \item MotIV-pwm2
      \item TFBSTools-PFMatrix
      \item TFBSTools-PWMatrix
      \item TFBSTools-ICMatrix
      \item seqLogo-pwm
      \item motifStack-pcm
      \item motifStack-pfm
      \item PWMEnrich-PWM
      \item Biostrings-PWM (type = "log2prob")
      \item rGADEM-motif
   }
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Convert a list of motifs.

\item \code{universalmotif}: Convert a \linkS4class{universalmotif} object.

\item \code{ANY}: Convert non-\linkS4class{universalmotif} class motifs.
}}

\examples{
# convert from universalmotif:
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.motifstack.pfm <- convert_motifs(jaspar, "motifStack-pfm")

# convert from another class to universalmotif:
library(TFBSTools)
data(MA0003.2)
motif <- convert_motifs(MA0003.2)

# convert from another class to another class
motif <- convert_motifs(MA0003.2, "PWMEnrich-PWM")

# the 'convert_motifs' function is embedded in the rest of the universalmotif
# functions; non-universalmotif class motifs can be used
MA0003.2.trimmed <- trim_motifs(MA0003.2)
# note: if the motif object going in has information that the
# 'universalmotif' class can't hold, it will be lost

}
\references{
\insertRef{motiv}{universalmotif}

   \insertRef{motifdb}{universalmotif}

   \insertRef{tfbstools}{universalmotif}

   \insertRef{seqlogo}{universalmotif}

   \insertRef{motifstack}{universalmotif}

   \insertRef{pwmenrich}{universalmotif}

   \insertRef{motifrg}{universalmotif}

   \insertRef{rgadem}{universalmotif}

   \insertRef{biostrings}{universalmotif}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
