% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_type.R
\name{convert_type}
\alias{convert_type}
\title{Convert \linkS4class{universalmotif} type.}
\usage{
convert_type(motif, type, pseudoweight, bkg, IC_floor = TRUE,
  IC_ceiling = TRUE)
}
\arguments{
\item{motif}{Motif object.}

\item{type}{Character. Either PCM, PPM, PWM, ICM.}

\item{pseudoweight}{Numeric.}

\item{bkg}{Numeric.}

\item{IC_floor}{Logical.}

\item{IC_ceiling}{Logical.}
}
\value{
Motif object.
}
\description{
Convert \linkS4class{universalmotif} type.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
