% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{convert_type}
\alias{convert_type}
\title{[UNDER CONSTRUCTION] Convert \linkS4class{universalmotif} type.}
\usage{
convert_type(motif, out_type, ...)
}
\arguments{
\item{motif}{Motif object.}

\item{out_type}{Character. Either PCM, PPM, PWM or ICM.}
}
\value{
Motif object of class \linkS4class{universalmotif}.
}
\description{
Convert between PCM, PPM, PWM and ICM. Cannot convert back from ICM.
}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
