% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_type.R
\name{convert_type}
\alias{convert_type}
\title{Convert \linkS4class{universalmotif} type.}
\usage{
convert_type(motifs, type, pseudoweight, bkg, IC_floor = TRUE,
  BPPARAM = bpparam())
}
\arguments{
\item{motifs}{Motif object or list.}

\item{type}{Character. Either PCM, PPM, PWM, ICM.}

\item{pseudoweight}{Numeric.}

\item{bkg}{Numeric.}

\item{IC_floor}{Logical.}

\item{BPPARAM}{Param for bplapply.}
}
\value{
Motif object.
}
\description{
Convert \linkS4class{universalmotif} type.
}
\examples{
jaspar.ppm <- read_jaspar(system.file("extdata", "jaspar.txt",
                                      package = "universalmotif"))
jaspar.pwm <- convert_type(jaspar.ppm, type = "PWM")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
