% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_type.R
\name{convert_type}
\alias{convert_type}
\title{Convert \linkS4class{universalmotif} type.}
\usage{
convert_type(
  motifs,
  type,
  pseudocount,
  nsize_correction = FALSE,
  relative_entropy = FALSE
)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable formats.}

\item{type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}.}

\item{pseudocount}{\code{numeric(1)} Correction to be applied to prevent \code{-Inf}
from appearing in PWM matrices. If missing, the pseudocount stored in the
\linkS4class{universalmotif} 'pseudocount' slot will be
used, which defaults to 0.}

\item{nsize_correction}{\code{logical(1)} If true, the ICM
at each position will be corrected to account
for small sample sizes. Only used if \code{relative_entropy = FALSE}.}

\item{relative_entropy}{\code{logical(1)} If true, the ICM will be
calculated as relative entropy. See details.}
}
\value{
See \code{\link[=convert_motifs]{convert_motifs()}} for possible output motif objects.
}
\description{
Switch between position count matrix (PCM), position probability matrix
(PPM), position weight matrix (PWM), and information count matrix (ICM)
types. See the "Introduction to sequence motifs" vignette for details.
}
\details{
Position count matrix (PCM), also known as position frequency matrix
(PFM). For n sequences from which the motif was built, each position is
represented by the numbers of each letter at that position. In theory
all positions should have sums equal to n, but not all databases are
this consistent. If converting from another type to PCM, column sums
will be equal to the 'nsites' slot; if empty, 100 is used.

Position probability matrix (PPM), also known as position frequency
matrix (PFM). At each position, the probability of individual letters
is calculated by dividing the count for that letter by the total sum of
counts at that position (\code{letter_count / position_total}).
As a result, each position will sum to 1. Letters with counts of 0 will
thus have a probability of 0, which can be undesirable when searching for
motifs in a set of sequences. To avoid this a pseudocount can be added
(\code{(letter_count + pseudocount) / (position_total + pseudocount)}).

Position weight matrix (PWM; \insertCite{pwm;textual}{universalmotif}),
also known as position-specific weight
matrix (PSWM), position-specific scoring matrix (PSSM), or
log-odds matrix. At each position, each letter is represented by it's
log-likelihood (\code{log2(letter_probability / background_probility)}),
which is normalized using the background letter frequencies. A PWM matrix
is constructed from a PPM; if any position has 0-probability letters to
which pseudocounts were not added, then the final log-likelihood of these
letters will be \code{-Inf}.

Information content matrix (ICM; \insertCite{icm;textual}{universalmotif}).
An ICM is a PPM where each letter probability is multiplied by the total
information content at that position. The information content of each
position is determined as: \code{totalIC - Hi}, where the total information
totalIC

\code{totalIC <- log2(alphabet_length)}, and the Shannon entropy
\insertCite{shannon}{universalmotif} for a specific
position (Hi)

\verb{Hi <- -sum(sapply(alphabet_frequencies, function(x) x * log(2))}.

As a result, the total sum or height of each position is representative of
it's sequence conservation, measured in the unit 'bits', which is a unit
of energy (\insertCite{bits;textual}{universalmotif}; see
\url{https://fr-s-schneider.ncifcrf.gov/logorecommendations.html}
for more information). However not all programs will calculate
information content the same; some will 'correct' the total information
content at each position using a correction factor as described by
\insertCite{correction;textual}{universalmotif}. This correction can
applied by setting \code{nsize_correction = TRUE}, however it will only
be applied if the 'nsites' slot is not empty. This is done using
\code{TFBSTools:::schneider_correction}
\insertCite{tfbstools}{universalmotif}. As such, converting from an ICM to
which some form of correction has been applied will result in a
PCM/PPM/PWM with slight inaccuracies.

Another method of calculating information content is calculating the
relative entropy, also known as Kullback-Leibler divergence
\insertCite{kl}{universalmotif}. This accounts for background
frequencies, which
can be useful for genomes with a heavy imbalance in letter frequencies.
For each position, the individual letter frequencies are calculated as
\code{letter_freq * log2(letter_freq / bkg_freq)}. When calculating
information content using Shannon entropy, the maximum content for
each position will always be \code{log2(alphabet_length)}; this does
not hold for information content calculated as relative entropy.
Please note that conversion from ICM assumes the information content
was \emph{not} calculated as relative entropy.
}
\examples{
jaspar.pcm <- read_jaspar(system.file("extdata", "jaspar.txt",
                                      package = "universalmotif"))

## pseudocounts default to 0.8
jaspar.pwm <- convert_type(jaspar.pcm, type = "PPM")

## setting pseudocounts to 0 will prevent any correction from being
## applied to PPM/PWM matrices
jaspar.pwm <- convert_type(jaspar.pcm, type = "PWM", pseudocount = 0)

}
\references{
\insertRef{kl}{universalmotif}

\insertRef{pseudo}{universalmotif}

\insertRef{correction}{universalmotif}

\insertRef{icm}{universalmotif}

\insertRef{bits}{universalmotif}

\insertRef{shannon}{universalmotif}

\insertRef{pwm}{universalmotif}

\insertRef{tfbstools}{universalmotif}
}
\seealso{
\code{\link[=convert_motifs]{convert_motifs()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
