% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{create_motif}
\alias{create_motif}
\alias{create_motif,character-method}
\alias{create_motif,matrix-method}
\alias{create_motif,XStringSet-method}
\title{Create a motif.}
\usage{
create_motif(input, name = "motif", pseudoweight = 0.8, alphabet, bkg,
  nsites, altname, family, organism, bkgsites, strand, pval, qval, eval,
  extrainfo)

\S4method{create_motif}{character}(input, name = "motif",
  pseudoweight = 0.8, alphabet, bkg, nsites, altname, family, organism,
  bkgsites, strand, pval, qval, eval, extrainfo)

\S4method{create_motif}{matrix}(input, name = "motif", pseudoweight = 0.8,
  alphabet, bkg, nsites, altname, family, organism, bkgsites, strand, pval,
  qval, eval, extrainfo)

\S4method{create_motif}{XStringSet}(input, name = "motif",
  pseudoweight = 0.8, alphabet, bkg, nsites, altname, family, organism,
  bkgsites, strand, pval, qval, eval, extrainfo)
}
\arguments{
\item{input}{One of character vector, matrix (PCM, PPM, or PWM), or 
XStringSet.}

\item{name}{Character.}

\item{pseudoweight}{Numeric.}

\item{alphabet}{Character.}

\item{bkg}{Numeric.}

\item{nsites}{Numeric.}

\item{altname}{Character.}

\item{family}{Character.}

\item{organism}{Character.}

\item{bkgsites}{Numeric.}

\item{strand}{Character.}

\item{pval}{Numeric.}

\item{qval}{Numeric.}

\item{eval}{Numeric.}

\item{extrainfo}{Character.}
}
\value{
Motif object.
}
\description{
Create a motif.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Create motif from a consensus string.

\item \code{matrix}: Create motif from a matrix.

\item \code{XStringSet}: Create motif from an XStringSet object.
}}

\examples{
my.motif <- create_motif("TATAA")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
