% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{create_motif}
\alias{create_motif}
\alias{create_motif,character,missing,missing-method}
\alias{create_motif,missing,matrix,missing-method}
\alias{create_motif,missing,missing,XStringSet-method}
\title{Create a motif.}
\usage{
create_motif(consensus, matrix, sequences, name = "motif",
  pseudoweight = 0.8, alphabet, bkg, nsites)

\S4method{create_motif}{character,missing,missing}(consensus, name,
  pseudoweight, alphabet, bkg, nsites)

\S4method{create_motif}{missing,matrix,missing}(matrix, name, pseudoweight,
  alphabet, bkg, nsites)

\S4method{create_motif}{missing,missing,XStringSet}(sequences, name,
  pseudoweight, alphabet, bkg, nsites)
}
\arguments{
\item{consensus}{Character. Consensus string.}

\item{matrix}{Matrix.}

\item{sequences}{XStringSet.}

\item{name}{Character.}

\item{pseudoweight}{Numeric.}

\item{alphabet}{Character.}

\item{bkg}{Numeric.}

\item{nsites}{Numeric.}
}
\value{
Motif object.
}
\description{
Create a motif.
}
\section{Methods (by class)}{
\itemize{
\item \code{consensus = character,matrix = missing,sequences = missing}: Create motif from a consensus string.

\item \code{consensus = missing,matrix = matrix,sequences = missing}: Create motif from a matrix.

\item \code{consensus = missing,matrix = missing,sequences = XStringSet}: Create motif from an XStringSet object.
}}

\examples{
my.motifs <- create_motif("TATAA")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
