% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{create_motif}
\alias{create_motif}
\alias{create_motif,character-method}
\alias{create_motif,matrix-method}
\alias{create_motif,XStringSet-method}
\title{Create a motif.}
\usage{
create_motif(input, name = "motif", pseudoweight = 0.8, alphabet, bkg,
  nsites)

\S4method{create_motif}{character}(input, name = "motif",
  pseudoweight = 0.8, alphabet, bkg, nsites)

\S4method{create_motif}{matrix}(input, name = "motif", pseudoweight = 0.8,
  alphabet, bkg, nsites)

\S4method{create_motif}{XStringSet}(input, name = "motif",
  pseudoweight = 0.8, alphabet, bkg, nsites)
}
\arguments{
\item{input}{One of character vector, matrix, or XStringSet.}

\item{name}{Character.}

\item{pseudoweight}{Numeric.}

\item{alphabet}{Character.}

\item{bkg}{Numeric.}

\item{nsites}{Numeric.}
}
\value{
Motif object.
}
\description{
Create a motif.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Create motif from a consensus string.

\item \code{matrix}: Create motif from a matrix.

\item \code{XStringSet}: Create motif from an XStringSet object.
}}

\examples{
my.motifs <- create_motif("TATAA")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
