% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_motif.R
\docType{methods}
\name{create_motif}
\alias{create_motif}
\alias{create_motif,missing-method}
\alias{create_motif,numeric-method}
\alias{create_motif,character-method}
\alias{create_motif,matrix-method}
\alias{create_motif,DNAStringSet-method}
\alias{create_motif,RNAStringSet-method}
\alias{create_motif,AAStringSet-method}
\alias{create_motif,BStringSet-method}
\title{Create a motif.}
\usage{
create_motif(input, alphabet, type = "PPM", name = "motif",
  pseudocount = 0, bkg, nsites, altname, family, organism, bkgsites,
  strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{missing}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{numeric}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{character}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{matrix}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{DNAStringSet}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{RNAStringSet}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{AAStringSet}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)

\S4method{create_motif}{BStringSet}(input, alphabet, type = "PPM",
  name = "motif", pseudocount = 0, bkg, nsites, altname, family,
  organism, bkgsites, strand, pval, qval, eval, extrainfo, add.multifreq)
}
\arguments{
\item{input}{\code{character}, \code{numberic}, \code{matrix},
\link[Biostrings:XStringSet-class]{Biostrings::XStringSet}, or \code{missing}}

\item{alphabet}{\code{character(1)} One of \code{c('DNA', 'RNA', 'AA', 'custom')},
or a combined string representing the letters.}

\item{type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}.}

\item{name}{\code{character(1)} Motif name.}

\item{pseudocount}{\code{numeric(1)} Correction to be applied to prevent \code{-Inf}
from apearing in PWM matrices.}

\item{bkg}{\code{numeric} Must sum to 1 and be equal in length to the alphabet
length.}

\item{nsites}{\code{numeric(1)} Number of sites the motif was constructed from.}

\item{altname}{\code{character(1)} Alternate motif name.}

\item{family}{\code{character(1)} Transcription factor family.}

\item{organism}{\code{character(1)} Species of origin.}

\item{bkgsites}{\code{numeric(1)} Total number of sites used to find the motif.}

\item{strand}{\code{character(1)} Whether the motif is specific to a certain strand.}

\item{pval}{\code{numeric(1)} P-value associated with motif.}

\item{qval}{\code{numeric(1)} Adjusted P-value associated with motif.}

\item{eval}{\code{numeric(1)} E-value associated with motif.}

\item{extrainfo}{\code{character} Any other extra information, represented as
a named character vector.}

\item{add.multifreq}{\code{numeric(1)} If the motif is created from a set of
sequences, then the \code{\link[=add_multifreq]{add_multifreq()}} function can be
run at the same type.}
}
\value{
\linkS4class{universalmotif} object.
}
\description{
Create a motif from a set of sequences, a matrix, or generate a random
motif.
}
\details{
The aim of this function is provide an easy interface to creating
\linkS4class{universalmotif} motifs, as an alternative to the
default class constructor (i.e. \code{new('universalmotif', name=...)}).
See examples for potential use cases.

Note: when generating random motifs, the \code{nsites} slot is also given a
random value. Furthermore, be careful about the \code{nsites} slot when creating
motifs from consensus strings: for example, the following call
\code{create_motif("TAAAT")} generates a motif with \code{nsites = 1}.

See the \code{examples} section for more info on motif creation.
}
\section{Methods (by class)}{
\itemize{
\item \code{missing}: Create a random motif of length 10.

\item \code{numeric}: Create a random motif with a specified length.

\item \code{character}: Create motif from a consensus string.

\item \code{matrix}: Create motif from a matrix.

\item \code{DNAStringSet}: Create motif from a \link[Biostrings:DNAStringSet-class]{Biostrings::DNAStringSet}.

\item \code{RNAStringSet}: Create motif from a \link[Biostrings:RNAStringSet-class]{Biostrings::RNAStringSet}.

\item \code{AAStringSet}: Create motif from a \link[Biostrings:AAStringSet-class]{Biostrings::AAStringSet}.

\item \code{BStringSet}: Create motif from a \link[Biostrings:BStringSet-class]{Biostrings::BStringSet}.
}}

\examples{
##### create motifs from a single string

# motif is by default generated as a PPM; change final type as desired
DNA.motif <- create_motif("TATAWAW")
DNA.motif <- create_motif("TATAWAW", type = "PCM")

# nsites will be set to the number of input sequences unless specified 
DNA.motif <- create_motif("TTTTTTT", nsites = 10)

# if ambiguity letters are found and nsites is not specified, nsites will
# be set to the minimum required to respect amibiguity letters
DNA.motif <- create_motif("TATAWAW")
DNA.motif <- create_motif("NNVVWWAAWWDDN")

# be careful about setting nsites when using ambiguity letters!
DNA.motif <- create_motif("NNVVWWAAWWDDN", nsites = 1)

RNA.motif <- create_motif("UUUCCG")

# 'create_motif' will try to detect the alphabet type; this can be 
# unreliable for AA and custom alphabets as DNA and RNA alphabets are
# detected first
AA.motif <- create_motif("AVLK", alphabet = "AA")

custom.motif <- create_motif("QWER", alphabet = "custom")
# specify custom alphabet
custom.motif <- create_motif("QWER", alphabet = "QWERASDF")

###### create motifs from multiple strings of equal length

DNA.motif <- create_motif(c("TTTT", "AAAA", "AACC", "TTGG"), type = "PPM")
DNA.motif <- create_motif(c("TTTT", "AAAA", "AACC", "TTGG"), nsites = 20)
RNA.motif <- create_motif(c("UUUU", "AAAA", "AACC", "UUGG"), type = "PWM")
AA.motif <- create_motif(c("ARNDCQ", "EGHILK", "ARNDCQ"), alphabet = "AA")
custom.motif <- create_motif(c("POIU", "LKJH", "POIU", "CVBN"),
                             alphabet = "custom")

# ambiguity letters are only allowed for single consensus strings; the
# following fails
\dontrun{
create_motif(c("WWTT", "CCGG"))
create_motif(c("XXXX", "XXXX"), alphabet = "AA")
}

##### create motifs from XStringSet objects

library(Biostrings)

DNA.set <- DNAStringSet(c("TTTT", "AAAA", "AACC", "TTGG"))
DNA.motif <- create_motif(DNA.set)
RNA.set <- RNAStringSet(c("UUUU", "AACC", "UUCC"))
RNA.motif <- create_motif(RNA.set)
AA.set <- AAStringSet(c("VVVLLL", "AAAIII"))
AA.motif <- create_motif(AA.set)

# custom motifs can be created from BStringSet objects
B.set <- BStringSet(c("QWER", "ASDF", "ZXCV", "TYUI"))
custom.motif <- create_motif(B.set)

##### create motifs with filled 'multifreq' slot

DNA.motif.k2 <- create_motif(DNA.set, add.multifreq = 2)

##### create motifs from matrices

mat <- matrix(c(1, 1, 1, 1,
                2, 0, 2, 0,
                0, 2, 0, 2,
                0, 0, 0, 0),
                nrow = 4, byrow = TRUE)
DNA.motif <- create_motif(mat, alphabet = "DNA")
RNA.motif <- create_motif(mat, alphabet = "RNA", nsites = 20)
custom.motif <- create_motif(mat)

# specify custom alphabet
custom.motif <- create_motif(mat, alphabet = "QWER")

# alphabet can be detected from rownames
rownames(mat) <- DNA_BASES
DNA.motif <- create_motif(mat)
rownames(mat) <- c("Q", "W", "E", "R")
custom.motif <- create_motif(mat)

# matrices can also be used as input
mat.ppm <- matrix(c(0.1, 0.1, 0.1, 0.1,
                    0.5, 0.5, 0.5, 0.5,
                    0.1, 0.1, 0.1, 0.1,
                    0.3, 0.3, 0.3, 0.3),
                    nrow = 4, byrow = TRUE)

DNA.motif <- create_motif(mat.ppm, alphabet = "DNA", type = "PPM")

##### create random motifs

# these are generated as PPMs with 10 positions

DNA.motif <- create_motif()
RNA.motif <- create_motif(alphabet = "RNA")
AA.motif <- create_motif(alphabet = "AA")
custom.motif <- create_motif(alphabet = "QWER")

# the number of positions can be specified

DNA.motif <- create_motif(5)

# If the background frequencies are not provided, they are generated
# using `rpois`; positions are created using `rdirichlet(1, bkg)`.
# (calling `create_motif()` creates motifs with an average
# positional IC of 1)

DNA.motif <- create_motif(bkg = c(0.3, 0.2, 0.2, 0.3))
DNA.motif <- create_motif(10, bkg = c(0.1, 0.4, 0.4, 0.1))

}
\seealso{
\code{\link[=convert_type]{convert_type()}}, \code{\link[=add_multifreq]{add_multifreq()}}, \code{\link[=create_sequences]{create_sequences()}},
\code{\link[=shuffle_motifs]{shuffle_motifs()}}.

\code{\link[=create_sequences]{create_sequences()}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
