% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sequences.R
\name{create_sequences}
\alias{create_sequences}
\title{Create random sequences.}
\usage{
create_sequences(alphabet = "DNA", seqnum = 100, seqlen = 100,
  monofreqs, difreqs, trifreqs)
}
\arguments{
\item{alphabet}{\code{character(1)} One of \code{c('DNA', 'RNA', 'AA')}, or a string of
letters/characters to be used as the alphabet.}

\item{seqnum}{\code{numeric(1)} Number of sequences to generate.}

\item{seqlen}{\code{numeric(1)} Length of random sequences.}

\item{monofreqs}{\code{numeric} Alphabet frequencies to use. If missing assumes uniform
frequencies. Not used if \code{difreq} or \code{trifreq} are
input.}

\item{difreqs}{\code{numeric} Dinucleotide frequencies. DNA/RNA only. Must be a
named numeric vector of length 16.}

\item{trifreqs}{\code{numeric} Trinucleotide frequencies. DNA/RNA only. Must be a 
named numeric vector of length 64.}
}
\value{
\linkS4class{XStringSet}
}
\description{
Generate random sequences from any set of characters.
}
\examples{
## create DNA sequences with slightly increased AT content:
sequences <- create_sequences(monofreqs = c(0.3, 0.2, 0.2, 0.3))
## create custom sequences:
sequences.QWER <- create_sequences("QWER")
## you can include non-alphabet characters are well, even spaces:
sequences.custom <- create_sequences("!@#$ ")

}
\references{
\insertRef{biostrings}{universalmotif}
}
\seealso{
\code{\link{create_motif}}, \code{\link{shuffle_sequences}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
