% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sequences.R
\name{create_sequences}
\alias{create_sequences}
\title{Create random sequences.}
\usage{
create_sequences(alphabet = "DNA", seqnum = 100, seqlen = 100, freqs,
  monofreqs, difreqs, trifreqs, progress = FALSE, BP = FALSE)
}
\arguments{
\item{alphabet}{\code{character(1)} One of \code{c('DNA', 'RNA', 'AA')}, or a string of
characters to be used as the alphabet.}

\item{seqnum}{\code{numeric(1)} Number of sequences to generate.}

\item{seqlen}{\code{numeric(1)} Length of random sequences.}

\item{freqs}{\code{numeric} A named vector of probabilities. The length of the
vector must be the power of the number of letters in the sequence alphabet.}

\item{monofreqs}{\code{numeric} Deprecated. Use \code{freqs} instead.}

\item{difreqs}{\code{numeric} Deprecated. Use \code{freqs} instead.}

\item{trifreqs}{\code{numeric} Deprecated. Use \code{freqs} instead.}

\item{progress}{\code{logical(1)} Show progress. Not recommended if \code{BP = TRUE}.}

\item{BP}{\code{logical(1)} Allows the use of \pkg{BiocParallel} within
\code{\link[=create_sequences]{create_sequences()}}. See \code{\link[BiocParallel:register]{BiocParallel::register()}} to change the default
backend. Setting \code{BP = TRUE} is only recommended for large jobs (such as
\code{create_sequences(seqlen=100000,seqnum=100000)}). Furthermore,
the behaviour of \code{progress = TRUE} is
changed if \code{BP = TRUE}; the default \pkg{BiocParallel} progress bar will
be shown (which unfortunately is much less informative).}
}
\value{
\code{\link{XStringSet}} The returned sequences are \emph{unnamed}.
}
\description{
Generate random sequences from any set of characters, represented as
\code{\link{XStringSet}} objects.
}
\examples{
## create DNA sequences with slightly increased AT content:
sequences <- create_sequences(freqs = c(A=0.3, C=0.2, G=0.2, T=0.3))
## create custom sequences:
sequences.QWER <- create_sequences("QWER")
## you can include non-alphabet characters are well, even spaces:
sequences.custom <- create_sequences("!@#$ ")

}
\references{
\insertRef{biostrings}{universalmotif}
}
\seealso{
\code{\link[=create_motif]{create_motif()}}, \code{\link[=shuffle_sequences]{shuffle_sequences()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
