% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_motifs.R
\name{enrich_motifs}
\alias{enrich_motifs}
\title{Enrich for input motifs in a set of sequences.}
\usage{
enrich_motifs(motifs, sequences, bkg.sequences, search.mode = "hits",
  max.p = 0.001, max.q = 0.001, qval.method = "fdr",
  positional.test = "t.test", threshold = 0.6, threshold.type = "logodds",
  verbose = TRUE, RC = TRUE, use.freq = 1, shuffle.k = 1,
  shuffle.method = "linear", shuffle.leftovers = "asis",
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{Motifs.}

\item{sequences}{DNAStringSet.}

\item{bkg.sequences}{DNAStringSet.}

\item{search.mode}{Character.}

\item{max.p}{Numeric.}

\item{max.q}{Numeric.}

\item{qval.method}{Numeric.}

\item{positional.test}{Character.}

\item{threshold}{Numeric.}

\item{threshold.type}{Character. One of 'logodds' and 'pvalue'.}

\item{verbose}{Logical.}

\item{RC}{Logical.}

\item{use.freq}{Numeric.}

\item{shuffle.k}{Numeric.}

\item{shuffle.method}{Character.}

\item{shuffle.leftovers}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
A list of results.
}
\description{
DNA/RNA only.
}
\author{
Benjamin Tremblay \email{b2tremblay@uwaterloo.ca}
}
