% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_motifs.R
\name{enrich_motifs}
\alias{enrich_motifs}
\title{Enrich for input motifs in a set of sequences.}
\usage{
enrich_motifs(motifs, sequences, bkg.sequences, search.mode = "hits",
  max.p = 1e-05, max.q = 1e-05, max.e = 0.001, qval.method = "fdr",
  positional.test = "t.test", threshold = 1e-04,
  threshold.type = "pvalue", verbose = 1, RC = FALSE, use.freq = 1,
  shuffle.k = 2, shuffle.method = "linear",
  shuffle.leftovers = "asis", return.scan.results = FALSE)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable motif formats.}

\item{sequences}{\code{XStringSet} Alphabet should match motif.}

\item{bkg.sequences}{\code{XStringSet} Optional; if missing,
\code{\link{shuffle_sequences}} is used to create background sequences from
the input sequences.}

\item{search.mode}{\code{character(1)} One of \code{c('hits', 'positional', 'both')}.
See details.}

\item{max.p}{\code{numeric(1)} P-value threshold.}

\item{max.q}{\code{numeric(1)} Adjusted P-value threshold. This is only useful
if multiple motifs are being enriched for.}

\item{max.e}{\code{numeric(1)}. The E-value is calculated by multiplying the adjusted
P-value with the number of input motifs \insertCite{meme2}{universalmotif}.}

\item{qval.method}{\code{character(1)} See \code{\link[stats]{p.adjust}}.}

\item{positional.test}{\code{character(1)} One of \code{c('t.test', 'wilcox.test',
'chisq.test', 'shapiro.test')}. If using the Shapiro test for
normality, then only the input sequences are tested for positionality;
the background sequences are ignored. See \code{\link[stats]{t.test}},
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{chisq.test}},
\code{\link[stats]{shapiro.test}}.}

\item{threshold}{\code{numeric(1)} Between 1 and 0. See \code{\link{scan_sequences}}.}

\item{threshold.type}{\code{character(1)} One of \code{c('logodds', 'pvalue')}. See
\code{\link{scan_sequences}}.}

\item{verbose}{\code{numeric(1)} 0 for no output, 4 for max verbosity.}

\item{RC}{\code{logical(1)} Whether to consider the reverse complement of the
sequences. Only available for \code{DNAStringSet}, \code{RNAStringSet}
sequences.}

\item{use.freq}{\code{numeric(1)} If the \code{multifreq} slot of the motifs are filled,
then they can be used to scan the sequences. See
\code{\link{scan_sequences}}.}

\item{shuffle.k}{\code{numeric(1)} The k-let size to use when shuffling input
sequences. Only used if no background sequences are input. See
\code{\link{shuffle_sequences}}.}

\item{shuffle.method}{\code{character(1)} See \code{\link{shuffle_sequences}}.}

\item{shuffle.leftovers}{\code{character(1)} See \code{\link{shuffle_sequences}}.}

\item{return.scan.results}{\code{logical(1)} Return output from
\code{\link{scan_sequences}}.}
}
\value{
\code{data.frame} Motif enrichment results. The resulting 
   \code{data.frame} contains the following columns:
   \itemize{
      \item \code{motif} Motif name.
      \item \code{total.seq.hits} Total number of matches accross all target
         sequences.
      \item \code{num.seqs.hits} Number of target sequences which contain matches.
      \item \code{num.seqs.total} Number of target sequences.
      \item \code{total.bkg.hits} Total number of matches accross all background
         sequences.
      \item \code{num.bkg.hits} Number of background sequences which contain
         matches.
      \item \code{num.bkg.total} Number of background sequences.
      \item \code{Pval.hits} P-value of enrichment. Only shown if
         \code{search.mode = c('hits', 'both')}.
      \item \code{Qval.hits} Q-val of enrichment. Only shown if
         \code{search.mode = c('hits', 'both')}.
      \item \code{Eval.hits} E-val of enrichment. Only shown if
         \code{search.mode = c('hits', 'both')}.
      \item \code{Pval.pos} P-value of positional comparison. Only
         shown if \code{search.mode = c('positional', 'both'}).
      \item \code{Qval.pos} Q-value of positional comparison. Only
         shown if \code{search.mode = c('positional', 'both'}).
      \item \code{Eval.pos} E-value of positional comparison. Only
         shown if \code{search.mode = c('positional', 'both'}).
   }
}
\description{
Given a set of target and background sequences, test if the input motifs
are significantly enriched in the targets sequences relative to the
background sequences.
}
\details{
To find enriched motifs, \code{\link{scan_sequences}} is run on both 
target and background sequences. If \code{search.mode = 'hits'},
\code{\link[stats]{fisher.test}} is run to test for enrichment. If
\code{search.mode = 'positional'}, then the test as set by
\code{positional.test} is run to check for positional differences
between target and background sequences. However if
\code{positional.test = 'shapiro.test'}, then only target sequence
hits are considered.
}
\examples{
target.sequences <- create_sequences(monofreqs = c(0.7, 0.1, 0.1, 0.1))
bkg.sequences <- create_sequences()
motif <- create_motif(bkg = c(0.7, 0.1, 0.1, 0.1))
enrich_motifs(motif, target.sequences, bkg.sequences)

}
\references{
\insertRef{meme2}{universalmotif}
}
\seealso{
\code{\link{scan_sequences}}, \code{\link{shuffle_sequences}},
   \code{\link{add_multifreq}}, \code{\link{motif_pvalue}}
}
\author{
Benjamin Tremblay \email{b2tremblay@uwaterloo.ca}
}
