% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_motifs.R
\name{enrich_motifs}
\alias{enrich_motifs}
\title{Enrich for input motifs in a set of sequences.}
\usage{
enrich_motifs(motifs, sequences, bkg.sequences, search.mode = "hits",
  max.p = 1e-05, max.q = 1e-05, max.e = 0.001, qval.method = "fdr",
  positional.test = "t.test", threshold = 0.001,
  threshold.type = "pvalue", verbose = 1, RC = FALSE, use.freq = 1,
  shuffle.k = 2, shuffle.method = "linear",
  shuffle.leftovers = "asis", return.scan.results = FALSE,
  progress = TRUE, BP = FALSE)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{sequences}{\code{\link{XStringSet}} Alphabet should match motif.}

\item{bkg.sequences}{\code{\link{XStringSet}} Optional; if missing,
\code{\link[=shuffle_sequences]{shuffle_sequences()}} is used to create background sequences from
the input sequences.}

\item{search.mode}{\code{character(1)} One of \code{c('hits', 'positional', 'both')}.
See details.}

\item{max.p}{\code{numeric(1)} P-value threshold.}

\item{max.q}{\code{numeric(1)} Adjusted P-value threshold. This is only useful
if multiple motifs are being enriched for.}

\item{max.e}{\code{numeric(1)}. The E-value is calculated by multiplying the adjusted
P-value with the number of input motifs times two
\insertCite{meme2}{universalmotif}.}

\item{qval.method}{\code{character(1)} See \code{\link[stats:p.adjust]{stats::p.adjust()}}.}

\item{positional.test}{\code{character(1)} One of \code{c('t.test', 'wilcox.test', 'chisq.test', 'shapiro.test')}. If using the Shapiro test for
normality, then only the input sequences are tested for positionality;
the background sequences are ignored. See \code{\link[stats:t.test]{stats::t.test()}},
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}, \code{\link[stats:chisq.test]{stats::chisq.test()}}, \code{\link[stats:shapiro.test]{stats::shapiro.test()}}.}

\item{threshold}{\code{numeric(1)} Between 1 and 0. See \code{\link[=scan_sequences]{scan_sequences()}}.}

\item{threshold.type}{\code{character(1)} One of \code{c('logodds', 'pvalue')}. See
\code{\link[=scan_sequences]{scan_sequences()}}.}

\item{verbose}{\code{numeric(1)} 0 for no output, 4 for max verbosity.}

\item{RC}{\code{logical(1)} Whether to consider the reverse complement of the
sequences. Only available for \code{\link{DNAStringSet}},
\code{\link{RNAStringSet}} sequences.}

\item{use.freq}{\code{numeric(1)} If the \code{multifreq} slot of the motifs are filled,
then they can be used to scan the sequences. See \code{\link[=scan_sequences]{scan_sequences()}}.}

\item{shuffle.k}{\code{numeric(1)} The k-let size to use when shuffling input
sequences. Only used if no background sequences are input. See
\code{\link[=shuffle_sequences]{shuffle_sequences()}}.}

\item{shuffle.method}{\code{character(1)} One of \code{c('markov', 'linear', 'random')}.
See \code{\link[=shuffle_sequences]{shuffle_sequences()}}. The \code{'random'} method is deprecated and
will be removed in the next minor version.}

\item{shuffle.leftovers}{\code{character(1)} One of \code{c('asis', 'first', 'split', 'discard')}.
Only used if \code{shuffle.method = 'random'}. See \code{\link[=shuffle_sequences]{shuffle_sequences()}}.}

\item{return.scan.results}{\code{logical(1)} Return output from
\code{\link[=scan_sequences]{scan_sequences()}}. For large jobs, leaving this as
\code{FALSE} can save a small amount time by preventing construction of the complete
results \code{data.frame} from \code{\link[=scan_sequences]{scan_sequences()}}.}

\item{progress}{\code{logical(1)} Show progress. Note recommended if \code{BP = TRUE}.
Set to \code{FALSE} if \code{verbose = 0}}

\item{BP}{\code{logical(1)} Allows the use of \pkg{BiocParallel} within
\code{\link[=enrich_motifs]{enrich_motifs()}}. See \code{\link[BiocParallel:register]{BiocParallel::register()}} to change the default
backend. Setting \code{BP = TRUE} is only recommended for exceptionally large jobs
(be wary of memory usage however, as \code{\link[=enrich_motifs]{enrich_motifs()}} does not try and
limit itself in this regard).
Furthermore, the behaviour of \code{progress = TRUE} is
changed if \code{BP = TRUE}; the default \pkg{BiocParallel} progress bar will
be shown (which unfortunately is much less informative).}
}
\value{
\code{data.frame} Motif enrichment results. The resulting
\code{data.frame} contains the following columns:
* \code{motif} Motif name.
* \code{total.seq.hits} Total number of matches across all target
sequences.
* \code{num.seqs.hits} Number of target sequences which contain matches.
* \code{num.seqs.total} Number of target sequences.
* \code{total.bkg.hits} Total number of matches across all background
sequences.
* \code{num.bkg.hits} Number of background sequences which contain
matches.
* \code{num.bkg.total} Number of background sequences.
* \code{Pval.hits} P-value of enrichment. Only shown if
\code{search.mode = c('hits', 'both')}.
* \code{Qval.hits} Q-val of enrichment. Only shown if
\code{search.mode = c('hits', 'both')}.
* \code{Eval.hits} E-val of enrichment. Only shown if
\code{search.mode = c('hits', 'both')}.
* \code{Pval.pos} P-value of positional comparison. Only
shown if \code{search.mode = c('positional', 'both')}.
* \code{Qval.pos} Q-value of positional comparison. Only
shown if \code{search.mode = c('positional', 'both')}.
* \code{Eval.pos} E-value of positional comparison. Only
shown if \code{search.mode = c('positional', 'both')}.
}
\description{
Given a set of target and background sequences, test if the input motifs
are significantly enriched in the targets sequences relative to the
background sequences.
}
\details{
To find enriched motifs, \code{\link[=scan_sequences]{scan_sequences()}} is run on both
target and background sequences. If \code{search.mode = 'hits'},
\code{\link[stats:fisher.test]{stats::fisher.test()}} is run to test for enrichment. If
\code{search.mode = 'positional'}, then the test as set by
\code{positional.test} is run to check for positional differences
between target and background sequences. However if
\code{positional.test = 'shapiro.test'}, then only target sequence
hits are considered.
}
\examples{
data(ArabidopsisPromoters)
data(ArabidopsisMotif)
enrich_motifs(ArabidopsisMotif, ArabidopsisPromoters, threshold = 0.01)

}
\references{
\insertRef{meme2}{universalmotif}
}
\seealso{
\code{\link[=scan_sequences]{scan_sequences()}}, \code{\link[=shuffle_sequences]{shuffle_sequences()}},
\code{\link[=add_multifreq]{add_multifreq()}}, \code{\link[=motif_pvalue]{motif_pvalue()}}
}
\author{
Benjamin Jean-Marie Tremblay \email{b2tremblay@uwaterloo.ca}
}
