% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_motifs.R
\name{enrich_motifs}
\alias{enrich_motifs}
\title{Enrich for input motifs in a set of sequences.}
\usage{
enrich_motifs(motifs, sequences, bkg.sequences, mode = "hits",
  max.p = 0.001, max.q = 0.001, qval.method = "fdr",
  positional.test = "t.test", threshold = 0.6, verbose = TRUE,
  RC = FALSE, HMMorder = 0, shuffle.k = 1, shuffle.method = "linear",
  shuffle.leftovers = "asis", plot = FALSE, plot.top = 5,
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{motifs}{Motifs.}

\item{sequences}{DNAStringSet.}

\item{bkg.sequences}{DNAStringSet.}

\item{mode}{Character.}

\item{max.p}{Numeric.}

\item{max.q}{Numeric.}

\item{qval.method}{Numeric.}

\item{positional.test}{Character.}

\item{threshold}{Numeric.}

\item{RC}{Logical.}

\item{HMMorder}{Numeric.}

\item{shuffle.k}{Numeric.}

\item{shuffle.method}{Character.}

\item{shuffle.leftovers}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
A list of results.
}
\description{
Enrich for input motifs in a set of sequences.
}
\author{
Benjamin Tremblay \email{b2tremblay@uwaterloo.ca}
}
