% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{filter_motifs}
\alias{filter_motifs}
\title{Filter a list of \linkS4class{universalmotif} motifs.}
\usage{
filter_motifs(motifs, name, altname, family, organism, width, alphabet, type,
  icscore, nsites, strand, pval, qval, eval)
}
\arguments{
\item{motifs}{List of motifs.}

\item{name}{Keep motifs by names.}

\item{altname}{Keep motifs by altnames.}

\item{family}{Keep motifs by family.}

\item{organism}{By organism.}

\item{width}{Keep motifs with minimum width.}

\item{alphabet}{Keep motifs by alphabet.}

\item{type}{Keep motifs by type.}

\item{icscore}{Keep motifs with minimum total IC.}

\item{nsites}{Keep motifs with minimum number of target sites.}

\item{strand}{Keeps motifs by strand.}

\item{pval}{Keep motifs by p-value.}

\item{qval}{Keep motifs by q-value.}

\item{eval}{Keep motifs by e-val.}
}
\value{
List of motifs.
}
\description{
Filter a list of \linkS4class{universalmotif} motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
