% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/universalmotif-methods.R,
%   R/methods.R
\docType{methods}
\name{filter_motifs}
\alias{filter_motifs}
\alias{filter_motifs,universalmotif-method}
\alias{filter_motifs,list-method}
\title{[UNDER CONSTRUCTION] Filter motifs.}
\usage{
filter_motifs(motifs, ...)

\S4method{filter_motifs}{universalmotif}(motifs, min_width, alphabet, type,
  icscore, nsites, strand, pval, eval, bkgsites, extrachar, extranum)

\S4method{filter_motifs}{list}(motifs, ...)
}
\arguments{
\item{motifs}{List of motifs.}

\item{...}{Extra parameters passed on to class-specific methods.}
}
\value{
List of motifs.
}
\description{
Filter a list of motifs.
}
\section{Methods (by class)}{
\itemize{
\item \code{universalmotif}: Filter \linkS4class{universalmotif}.

\item \code{list}: Filter a list of motifs.
}}

\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
