% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{filter_motifs}
\alias{filter_motifs}
\title{Filter a list of \linkS4class{universalmotif} motifs.}
\usage{
filter_motifs(motifs, name, altname, family, organism, width, alphabet,
  type, icscore, nsites, strand, pval, qval, eval,
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{\code{list} See \code{\link{convert_motifs}} for acceptable
formats.}

\item{name}{\code{character} Keep motifs by names.}

\item{altname}{\code{altname} Keep motifs by altnames.}

\item{family}{\code{family} Keep motifs by family.}

\item{organism}{\code{organism} Keep motifs by organism.}

\item{width}{\code{numeric(1)} Keep motifs with minimum width.}

\item{alphabet}{\code{character} Keep motifs by alphabet.}

\item{type}{\code{character} Keep motifs by type.}

\item{icscore}{\code{numeric(1)} Keep motifs with minimum total IC.}

\item{nsites}{\code{numeric(1)} Keep motifs with minimum number of target sites.}

\item{strand}{\code{character} Keeps motifs by strand.}

\item{pval}{\code{numeric(1)} Keep motifs by max P-value.}

\item{qval}{\code{numeric(1)} Keep motifs by max Q-value.}

\item{eval}{\code{numeric(1)} Keep motifs by max E-val.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
\code{list} Motifs.
}
\description{
Filter motifs based on the contents of available \linkS4class{universalmotif}
slots.
}
\examples{
## By minimum IC:
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.ic10 <- filter_motifs(jaspar, icscore = 10)

## By organism:
if (requireNamespace("MotifDb", quietly = TRUE)) {
  library(MotifDb)
  motifs <- convert_motifs(MotifDb)
  motifs <- filter_motifs(motifs, organism = c("Athaliana", "Mmusculus"))
}

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
