% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motifs.R
\name{filter_motifs}
\alias{filter_motifs}
\title{Filter a list of \linkS4class{universalmotif} motifs.}
\usage{
filter_motifs(motifs, name, altname, family, organism, width, alphabet,
  type, icscore, nsites, strand, pval, qval, eval,
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs.}

\item{name}{Keep motifs by names.}

\item{altname}{Keep motifs by altnames.}

\item{family}{Keep motifs by family.}

\item{organism}{Keep motifs by organism.}

\item{width}{Keep motifs with minimum width.}

\item{alphabet}{Keep motifs by alphabet.}

\item{type}{Keep motifs by type.}

\item{icscore}{Keep motifs with minimum total IC.}

\item{nsites}{Keep motifs with minimum number of target sites.}

\item{strand}{Keeps motifs by strand.}

\item{pval}{Keep motifs by max P-value.}

\item{qval}{Keep motifs by max Q-value.}

\item{eval}{Keep motifs by max E-val.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
List of motifs.
}
\description{
Filter motifs based on the contents of available \linkS4class{universalmotif}
slots.
}
\examples{
# By minimum IC:
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.ic10 <- filter_motifs(jaspar, icscore = 10)

# By organism:
library(MotifDb)
motifs <- convert_motifs(MotifDb)
motifs <- filter_motifs(motifs, organism = c("Athaliana", "Mmusculus"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
