% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JASPAR2018_CORE_DBSCORES}
\alias{JASPAR2018_CORE_DBSCORES}
\title{JASPAR2018 CORE database scores.}
\format{List with three data.frames.}
\usage{
JASPAR2018_CORE_DBSCORES
}
\description{
For use with \code{\link{compare_motifs}}.
}
\keyword{datasets}
