% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DBscores.R
\name{make_DBscores}
\alias{make_DBscores}
\title{Create P-value databases.}
\usage{
make_DBscores(db.motifs, method = c("PCC", "EUCL", "SW", "KL", "ALLR",
  "BHAT", "HELL", "IS", "SEUCL", "MAN", "ALLR_LL"), shuffle.db = TRUE,
  shuffle.k = 3, shuffle.method = "linear", rand.tries = 1000,
  widths = 5:30, min.position.ic = 0, normalise.scores = c(FALSE,
  TRUE), min.overlap = 6, min.mean.ic = 0, progress = TRUE,
  BP = FALSE, nthreads = 1, tryRC = TRUE, score.strat = c("sum",
  "a.mean", "g.mean", "median"))
}
\arguments{
\item{db.motifs}{\code{list} Database motifs.}

\item{method}{\code{character(1)} One of PCC, EUCL, SW, KL, ALLR, BHAT, HELL, IS,
SEUCL, MAN, ALLR_LL. See details.}

\item{shuffle.db}{\code{logical(1)} Deprecated. Does nothing.
generate random motifs with \code{\link[=create_motif]{create_motif()}}.}

\item{shuffle.k}{\code{numeric(1)} See \code{\link[=shuffle_motifs]{shuffle_motifs()}}.}

\item{shuffle.method}{\code{character(1)} See \code{\link[=shuffle_motifs]{shuffle_motifs()}}.}

\item{rand.tries}{\code{numeric(1)} Approximate number of comparisons
to perform for every combination of \code{widths}.}

\item{widths}{\code{numeric} Motif widths to use in P-value database calculation.}

\item{min.position.ic}{\code{numeric(1)} Minimum information content required between
individual alignment positions for it to be counted in the final alignment
score. It is recommended to use this together with \code{normalise.scores = TRUE},
as this will help punish scores resulting from only a fraction of an
alignment.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions, as well as alignments between motifs of similar length.
Similarity scores are multiplied by the ratio of
aligned positions to the total number of positions in the larger motif,
and the inverse for distance scores.}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{\code{numeric(1)} Minimum mean information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{progress}{\code{logical(1)} Show progress.}

\item{BP}{\code{logical(1)} Deprecated. See \code{nthreads}.}

\item{nthreads}{\code{numeric(1)} Run \code{\link[=compare_motifs]{compare_motifs()}} in parallel with \code{nthreads}
threads. \code{nthreads = 0} uses all available threads.}

\item{tryRC}{\code{logical(1)} Try the reverse complement of the motifs as well,
report the best score.}

\item{score.strat}{\code{character(1)} How to handle column scores calculated from
motif alignments. "sum": add up all scores. "a.mean": take the arithmetic
mean. "g.mean": take the geometric mean. "median": take the median.}
}
\value{
A \code{DataFrame} with score distributions for the
input database. If more than one \code{\link[=make_DBscores]{make_DBscores()}} run occurs (i.e. args
\code{method}, \code{normalise.scores} or \code{score.strat} are longer than 1), then
a \code{list} is returned with a \code{DataFrame} of scores and a \code{list} of
function parameters.
}
\description{
Generate data used by \code{\link[=compare_motifs]{compare_motifs()}} for P-value calculations. By default,
\code{\link[=compare_motifs]{compare_motifs()}} uses an internal database based on the JASPAR2018 core motifs
\insertCite{jaspar}{universalmotif}. Parameters for a logistic distribution are
are estimated for every combination of motif \code{widths}.
}
\details{
See \code{\link[=compare_motifs]{compare_motifs()}} for more info on comparison parameters.

To replicate the internal \pkg{universalmotif} DB scores, run
\code{\link[=make_DBscores]{make_DBscores()}} with the default settings. Note that this will be
a slow process.

Arguments \code{widths}, \code{method}, \code{normalise.scores} and \code{score.strat} are
vectorized; all combinations will be attempted.

Randomly generated scores are used to estimate parameters for logistic
distributions. This will occasionally fail for \code{method = "IS"}, which
can sometimes fit too poorly to estimate parameters. In such cases,
the mean and standard deviation will instead be used. A warning is
given in \code{\link[=compare_motifs]{compare_motifs()}} when P-values are asked for when using
\code{method = "IS"}.
}
\examples{
\dontrun{
library(MotifDb)
motifs <- convert_motifs(MotifDb[1:100])
scores <- make_DBscores(motifs, method = "PCC")
compare_motifs(motifs, 1:100, db.scores = scores)
}

}
\references{
\insertRef{jaspar}{universalmotif}
}
\seealso{
\code{\link[=compare_motifs]{compare_motifs()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
