% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_motifs.R
\name{merge_motifs}
\alias{merge_motifs}
\title{Merge motifs using msa or motifStack}
\usage{
merge_motifs(motifs, newname = "merged motif", method = "msa",
  printaln = FALSE, tryRC = TRUE, RCstrategy = "motif_dist",
  bgNoise = NA, cluster = "default", substitutionMatrix = "iub",
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{motifs}{List of motifs.}

\item{newname}{Name of final merged motif.}

\item{method}{Character. 'motifStack' or 'msa'.}

\item{printaln}{Logical. Print msa aln.}

\item{tryRC}{Logical. Only used if method = 'msa'.}

\item{RCstrategy}{Character. 'motif_dist' or 'motif_simil'.}

\item{bgNoise}{\code{\link[motifStack]{mergeMotifs}} param.}

\item{cluster}{\code{\link[msa]{msaClustalW}} param. nj (default) or upgma}

\item{substitutionMatrix}{\code{\link[msa]{msaClustalW}} param. iub or clustalw}

\item{BPPARAM}{Param for bplapply. Only used if tryRC = TRUE.}

\item{...}{Other settings for function msaClustalW.}
}
\value{
A single motif object.
}
\description{
DNA/RNA only. The 'extrainfo' slot will be dropped, as well as any
non-identical slots.
}
\details{
If method = 'msa', the motifs are aligned as consensus strings before
   they are averaged as PPMs. For method = 'motifStack', the
   motifStack::mergeMotifs function is used. 

   If tryRC = FALSE, the motifs are aligned and merged as-is; if
   tryRC = TRUE, then either \code{\link{motif_simil}} or
   \code{\link{motif_dist}} is used to check if the reverse complement
   is more closely related to the rest of the motifs. If so, the
   reverse complement is used instead for merging.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
