% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_motifs.R
\name{merge_motifs}
\alias{merge_motifs}
\title{Merge motifs.}
\usage{
merge_motifs(motifs, method = "MPCC", use.type = "PPM",
  min.overlap = 6, min.mean.ic = 0.5, tryRC = TRUE,
  relative_entropy = FALSE, normalise.scores = FALSE)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{method}{\code{character(1)} One of \code{c('PCC', 'MPCC', 'EUCL', 'MEUCL', 'SW', 'MSW', 'KL', 'MKL')}. See details.}

\item{use.type}{\code{character(1) One of}'PPM'\code{and}'ICM'\code{. The latter allows for taking into account the background frequencies if}relative_entropy = TRUE`.}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{\code{numeric(1)} Minimum information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{tryRC}{\code{logical} Try the reverse complement of the motifs as well,
report the best score.}

\item{relative_entropy}{\code{logical(1)} For ICM calculation. See
\code{\link[=convert_type]{convert_type()}}.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions, as well as alignments between motifs of similar length.
Similarity scores are multiplied by the ratio of
aligned positions to the total number of positions in the larger motif,
and the inverse for distance scores.}
}
\value{
A single motif object. See \code{\link[=convert_motifs]{convert_motifs()}} for
available formats.
}
\description{
Aligns the motifs using \code{\link[=compare_motifs]{compare_motifs()}}, then averages the
motif PPMs. Currently the \code{multifreq} slot, if filled in any of the motifs,
will be dropped.
}
\examples{
if (requireNamespace("MotifDb", quietly = TRUE)) {
  library(MotifDb)
  merged.motif <- merge_motifs(MotifDb[1:5])
}

}
\seealso{
\code{\link[=compare_motifs]{compare_motifs()}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
