% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_motifs.R
\name{merge_motifs}
\alias{merge_motifs}
\title{Merge motifs.}
\usage{
merge_motifs(motifs, method = "Pearson", min.overlap = 6,
  min.mean.ic = 0.5, tryRC = TRUE, relative_entropy = FALSE,
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs.}

\item{method}{Character.}

\item{min.overlap}{Numeric.}

\item{min.mean.ic}{Numeric.}

\item{tryRC}{Logical.}

\item{relative_entropy}{Logical.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
A single motif object.
}
\description{
DNA/RNA only. The 'extrainfo' slot will be dropped, as well as any
non-identical slots.
}
\examples{
motifs <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))

merged.motif <- merge_motifs(motifs)
# compare with:
merged.motif2 <- merge_motifs(motifs, method = "motifStack")

}
\seealso{
\code{\link{compare_motifs}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
