% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_logo.R
\name{motif_logo}
\alias{motif_logo}
\title{Generate ggplot2 motif logos.}
\usage{
motif_logo(motif, engine = "ggseqlogo", type = "ic", BPPARAM = bpparam(),
  ...)
}
\arguments{
\item{motif}{universalmotif object.}

\item{engine}{Character. ggseqlogo, seqLogo, motifStack, Logolas.}

\item{type}{Character. For \code{engine = c('ggseqlogo', 'seqLogo',
'motifStack')}: 'ic' or 'prob'. For \code{engine = Logolas}:
'Logo' or 'EDLogo'.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{...}{Package-specific params.}
}
\value{
For ggseqlogo: a ggplot2 object.
}
\description{
Create logos using ggplot2 + ggseqlogo, seqLogo, motifStack or Logolas.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.logo <- motif_logo(jaspar[[1]], type = "prob")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
