% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_pvalue.R
\name{motif_pvalue}
\alias{motif_pvalue}
\title{Motif P-value and scoring utility}
\usage{
motif_pvalue(motif, score, pvalue, bkg.probs, k = 6, tolerance = 0.75)
}
\arguments{
\item{motif}{Single motif.}

\item{score}{Numeric.}

\item{pvalue}{Numeric.}

\item{bkg.probs}{Numeric.}

\item{k}{Numeric.}

\item{tolerance}{Numeric.}
}
\value{
Numeric.
}
\description{
Motif P-value and scoring utility
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
