% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_pvalue.R
\name{motif_pvalue}
\alias{motif_pvalue}
\title{Motif P-value and scoring utility}
\usage{
motif_pvalue(motifs, score, pvalue, bkg.probs, use.freq = 1, k = 6,
  progress = ifelse(length(motifs) > 1, TRUE, FALSE), BP = FALSE)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{score}{\code{numeric} Get a p-value for a motif from a logodds score.}

\item{pvalue}{\code{numeric} Get a logodds score for a motif from a
p-value.}

\item{bkg.probs}{\code{numeric}, \code{list} If supplying individual background
probabilities for each motif, a list. If missing, assumes a uniform
background. Note that this only influences calculating p-values
from an input score; calculating a score from an input p-value
currently assumes a uniform background.}

\item{use.freq}{\code{numeric(1)} By default uses the regular motif matrix;
otherwise uses the corresponding \code{multifreq} matrix.}

\item{k}{\code{numeric(1)} For speed, scores/p-values can be approximated after
subsetting the motif every \code{k} columns. If \code{k} is a value
equal or higher to the size of input motif(s), then the calculations
are (nearly) exact.}

\item{progress}{\code{logical(1)} Show progress. Not recommended if \code{BP = TRUE}.}

\item{BP}{\code{logical(1)} Allows the use of \pkg{BiocParallel} within
\code{\link[=motif_pvalue]{motif_pvalue()}}. See \code{\link[BiocParallel:register]{BiocParallel::register()}} to change the default
backend. Setting \code{BP = TRUE} is only recommended for exceptionally large
jobs. Furthermore, the behaviour of \code{progress = TRUE} is changed
if \code{BP = TRUE}; the default \pkg{BiocParallel} progress bar will be
shown (which unfortunately is much less informative).}
}
\value{
\code{numeric} A vector of scores/p-values.
}
\description{
For calculating p-values/logodds scores for any number of motifs.
}
\details{
Calculating p-values for motifs can be very computationally intensive. This
is due to how p-values must be calculated: for a given score, all possible
sequences which score equal or higher must be found, and the probability for
each of these sequences (based on background probabilities) summed. For a DNA
motif of length 10, the number of possible unique sequences is 4^10 = 1,048,576.
Finding all possible sequences higher than a given score can be done
very efficiently and quickly with a branch-and-bound algorithm, but as the
motif length increases this calculation becomes impractical. To get
around this, the p-value calculation can be approximated.

In order to calculate p-values for longer motifs, this function uses the
approximation proposed by \insertCite{pvalues;textual}{universalmotif}, where
the motif is subset, p-values calculated for the subsets, and finally
combined for a total p-value. The smaller the size of the subsets, the
faster the calculation; but also, the bigger the approximation. This can be
controlled by setting \code{k}. In fact, for smaller motifs (< 13 positions)
calculating exact p-values can be done individually in reasonable time by
setting \code{k = 12}.

To calculate a score based on a given p-value, the means and variances of
each motif subsets are combined to estimate the distribution of all
possible scores using \code{\link[stats:qnorm]{stats::qnorm()}}:

\code{qnorm(pvalue, mean = sum(subset.means), sd = sqrt(sum(subset.vars)))}

For calculating exact scores, \code{\link[stats:ecdf]{stats::ecdf()}} and \code{\link[stats:quantile]{stats::quantile()}} are
used:

\code{quantile(ecdf(scores), probs = pvalue)}

It is important to keep in mind that both approximate and exact score
calculations assume uniform backgrounds. To get all possible scores for
each subset, \code{\link[=expand.grid]{expand.grid()}} is used instead of the branch-and-bound
algorithm used for calculating p-values. Keep this in mind for determining
the best \code{k} value for motifs with alphabets longer than DNA/RNA motifs.
}
\examples{
data(examplemotif)

## p-value/score calculations are performed using the PWM version of the
## motif; these calculations do not work if any -Inf values are present
examplemotif["pseudocount"] <- 1
# or
examplemotif <- BiocGenerics::normalize(examplemotif)

## get a minimum score based on a p-value
motif_pvalue(examplemotif, pvalue = 0.001)

## get the probability of a particular sequence hit
motif_pvalue(examplemotif, score = 0)

## the calculations can be performed for multiple motifs
motif_pvalue(list(examplemotif, examplemotif), pvalue = c(0.001, 0.0001))

## get motif site p-values after using scan_sequences()
data(ArabidopsisMotif)
data(ArabidopsisPromoters)
res <- scan_sequences(ArabidopsisMotif, ArabidopsisPromoters, RC = FALSE,
                      progress = FALSE, verbose = 0, threshold = 0,
                      threshold.type = "logodds")[1:100, ]
res$pvalue <- motif_pvalue(ArabidopsisMotif, score = res$score)

}
\references{
\insertRef{pvalues}{universalmotif}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
