% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_rc.R
\name{motif_rc}
\alias{motif_rc}
\title{Get the reverse complement of a motif.}
\usage{
motif_rc(motifs)
}
\arguments{
\item{motifs}{List of motifs or a single motif.}
}
\description{
Get the reverse complement of a motif.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
