% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_simil.R
\name{motif_simil}
\alias{motif_simil}
\title{Motif similarity matrix.}
\usage{
motif_simil(motifs, method = "Pearson", BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs.}

\item{method}{One of Euclidean, Pearson, KL.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Similarity matrix.
}
\description{
TFBSTools implementation of motif similarity
\insertCite{tfbstools}{universalmotif}.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.simil <- motif_simil(jaspar)

}
\references{
\insertRef{tfbstools}{universalmotif}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
