% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_simil.R
\name{motif_simil}
\alias{motif_simil}
\title{Motif similarity matrix.}
\usage{
motif_simil(motifs, method = "Pearson")
}
\arguments{
\item{motifs}{List of motifs.}

\item{method}{One of Euclidean, Pearson, KL.}
}
\value{
Similarity matrix.
}
\description{
TFBSTools implementation of motif similarity. Supports parallel execution
via foreach package.
}
\examples{
\dontrun{
library(doParallel)
registerDoParallel()
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.simil <- motif_simil(jaspar)
}

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
