% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{motif_slots}
\alias{motif_slots}
\title{Accessor function for class universalmotif.}
\usage{
motif_slots(object, ...)
}
\arguments{
\item{object}{Motif object of class \linkS4class{universalmotif}.}
}
\value{
List of motif info.
}
\description{
Easy access to universalmotif slots.
}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
