% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_tree.R
\name{motif_tree}
\alias{motif_tree}
\title{Generate ggplot2 motif trees with ggtree.}
\usage{
motif_tree(motifs, layout = "circular", linecol = "family",
  labels = "none", tipsize = "none", legend = TRUE,
  branch.length = "none", db.scores, method = "Pearson",
  use.type = "PPM", min.overlap = 6, tryRC = TRUE,
  min.mean.ic = 0.5, relative_entropy = FALSE, ...)
}
\arguments{
\item{motifs}{\code{list}, \code{dist} See \code{\link{convert_motifs}} for
available formats.}

\item{layout}{\code{character(1)} One of \code{c('rectangular', 'slanted', 'fan', 'circular',
'radial', 'equal_angle', 'daylight')}.}

\item{linecol}{\code{character(1)} \linkS4class{universalmotif} slot to use to
colour lines (e.g. 'family').}

\item{labels}{\code{character(1)} \linkS4class{universalmotif} slot to use to label
tips (e.g. 'name').}

\item{tipsize}{\code{character(1)} \linkS4class{universalmotif} slot to use to
control tip size (e.g. 'icscore').}

\item{legend}{\code{logical(1)} Show legend for line colour and tip size.}

\item{branch.length}{\code{character(1)} If 'none', draw a cladogram.}

\item{db.scores}{\code{data.frame} See \code{\link{compare_motifs}}.}

\item{method}{\code{character(1)} One of \code{'Pearson', 'Euclidean', 'KL'}.}

\item{use.type}{\code{character(1)} One of \code{'PCM'} (Pearson only),
\code{'PPM'} (any method), \code{'PWM'} (Pearson only), and \code{'ICM'}
(any method). The two allow for taking into account the background
frequencies (for ICM, only if \code{relative_entropy = TRUE}).}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{tryRC}{\code{logical(1)} Try the reverse complement of the motifs as well,
report the best score.}

\item{min.mean.ic}{\code{numeric(1)} Minimum information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{relative_entropy}{\code{logical(1)} For ICM calculation. See
\code{\link{convert_type}}.}

\item{...}{\pkg{ggtree} params.}
}
\value{
\code{ggplot} object.
}
\description{
For more powerful motif tree functions, see the motifStack package.
}
\details{
See \code{\link[ggtree]{ggtree}} for more detailed descriptions of
   parameters.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.tree <- motif_tree(jaspar, linecol = "none", labels = "name",
                          layout = "rectangular")

}
\references{
\insertRef{ggplot2}{universalmotif}

   \insertRef{ggtree}{universalmotif}
}
\seealso{
\code{\link[motifStack]{motifStack}}, \code{\link{compare_motifs}},
   \code{\link[ggtree]{ggtree}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
