% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_tree.R
\name{motif_tree}
\alias{motif_tree}
\title{Generate ggplot2 motif trees with ggtree.}
\usage{
motif_tree(motifs, layout = "circular", linecol = "family",
  labels = "none", tipsize = "none", legend = TRUE,
  branch.length = "none", BPPARAM = bpparam(), ...)
}
\arguments{
\item{motifs}{List of motifs or 'dist' object.}

\item{layout}{Character.}

\item{linecol}{Character. (family, organism, etc)}

\item{labels}{Character. (name, altname, etc)}

\item{tipsize}{Character. (icscore, pval, qval, eval)}

\item{legend}{Logical.}

\item{branch.length}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{...}{ggtree params.}
}
\value{
ggplot2 object.
}
\description{
For more powerful motif tree functions, see motifStack.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.tree <- motif_tree(jaspar, linecol = "none", labels = "name",
                          layout = "rectangular")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
