% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_tree.R
\name{motif_tree}
\alias{motif_tree}
\title{Generate ggplot2 motif trees with ggtree.}
\usage{
motif_tree(motifs, layout = "circular", linecol = "family",
  labels = "none", tipsize = "none", legend = TRUE,
  branch.length = "none", db.scores, method = "Pearson",
  use.type = "PPM", min.overlap = 6, tryRC = TRUE,
  min.mean.ic = 0.5, relative_entropy = FALSE,
  BPPARAM = SerialParam(), ...)
}
\arguments{
\item{motifs}{List of \linkS4class{universalmotif} objects or 'dist' object.}

\item{layout}{Character. One of 'rectangular', 'slanted', 'fan', 'circular',
'radial', 'equal_angle', and 'daylight'.}

\item{linecol}{Character. \linkS4class{universalmotif} slot to use to
colour lines (e.g. 'family').}

\item{labels}{Character. \linkS4class{universalmotif} slot to use to label
tips (e.g. 'name').}

\item{tipsize}{Character. \linkS4class{universalmotif} slot to use to
control tip size (e.g. 'icscore').}

\item{legend}{Logical. Show legend for line colour and tip size.}

\item{branch.length}{Character. If 'none', draw a cladogram.}

\item{db.scores}{data.frame.}

\item{method}{Character.}

\item{use.type}{Character.}

\item{min.overlap}{Numeric.}

\item{tryRC}{Logical.}

\item{min.mean.ic}{Numeric.}

\item{relative_entropy}{Logical.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{...}{ggtree params.}
}
\value{
ggplot2 object.
}
\description{
For more powerful motif tree functions, see the motifStack package.
}
\details{
See \code{\link[ggtree]{ggtree}} for more detailed descriptions of
   parameters.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.tree <- motif_tree(jaspar, linecol = "none", labels = "name",
                          layout = "rectangular")

}
\references{
\insertRef{ggtree}{universalmotif}

   \insertRef{ggplot2}{universalmotif}
}
\seealso{
\code{\link[motifStack]{motifStack}}, \code{\link{compare_motifs}},
   \code{\link[ggtree]{ggtree}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
