% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cisbp.R
\name{read_cisbp}
\alias{read_cisbp}
\title{Import CIS-BP motifs.}
\usage{
read_cisbp(file, skip = 0, BPPARAM = SerialParam())
}
\arguments{
\item{file}{\code{character(1)} File name.}

\item{skip}{\code{numeric(1)} If not zero, will skip however many desired lines in the
file before starting to read.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
\code{list} \linkS4class{universalmotif} objects.
}
\description{
Import CIS-BP formatted motifs. See \url{http://cisbp.ccbr.utoronto.ca/index.php}.
Assumed to be DNA motifs.
}
\examples{
cisbp <- read_cisbp(system.file("extdata", "cisbp.txt",
                                package = "universalmotif"))

}
\references{
\insertRef{cisbp}{universalmotif}
}
\seealso{
Other read_motifs: \code{\link{read_homer}},
  \code{\link{read_jaspar}}, \code{\link{read_matrix}},
  \code{\link{read_meme}}, \code{\link{read_transfac}},
  \code{\link{read_uniprobe}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{read_motifs}
