% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cisbp.R
\name{read_cisbp}
\alias{read_cisbp}
\title{Import CIS-BP motifs.}
\usage{
read_cisbp(file, skip = 0)
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}
}
\description{
Import CIS-BP motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
