% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer.R
\name{read_homer}
\alias{read_homer}
\title{Import HOMER motifs.}
\usage{
read_homer(file, skip = 0)
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}
}
\description{
Assumed DNA.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
