% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer.R
\name{read_homer}
\alias{read_homer}
\title{Import HOMER motifs.}
\usage{
read_homer(file, skip = 0, BPPARAM = bpparam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
List of universalmotif objects.
}
\description{
Assumed DNA.
}
\examples{
homer <- read_homer(system.file("extdata", "homer.txt",
                                package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
