% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer.R
\name{read_homer}
\alias{read_homer}
\title{[UNDER CONSTRUCTION] Load Homer motifs from a text file.}
\usage{
read_homer(motif_file, verbose = FALSE, show_warnings = TRUE,
  mot_length_cutoff = NULL, bkg_cutoff = NULL, target_cutoff = NULL,
  log_odds_thresh = NULL, p_enrich_cutoff = NULL, p_detect_cutoff = NULL,
  tpos_cutoff = NULL, tstd_cutoff = NULL, bpos_cutoff = NULL,
  bstd_cutoff = NULL, strand_bias_cutoff = NULL,
  multiplicity_cutoff = NULL)
}
\arguments{
\item{motif_file}{Character.}

\item{verbose}{Logical.}

\item{show_warnings}{Logical.}

\item{mot_length_cutoff}{Integer.}

\item{bkg_cutoff}{Double. Percent. Motifs with more than num are cut.}

\item{target_cutoff}{Double. Percent. Motifs with more than num are kept.}

\item{log_odds_thresh}{Double.}

\item{p_enrich_cutoff}{Double.}

\item{p_detect_cutoff}{Double.}

\item{tpos_cutoff}{Double.}

\item{tstd_cutoff}{Double.}

\item{bpos_cutoff}{Double.}

\item{bstd_cutoff}{Double.}

\item{strand_bias_cutoff}{Double.}

\item{multiplicity_cutoff}{Double.}
}
\value{
A list of motifs.
}
\description{
PLACEHOLDER TEXT.
}
\examples{
  motifs <- system.file("extdata", "example.homer", package = "universalmotif")
  rmotifs <- read_homer(motifs)

}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
