% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer.R
\name{read_homer}
\alias{read_homer}
\title{Import HOMER motifs.}
\usage{
read_homer(file, skip = 0)
}
\arguments{
\item{file}{\code{character(1)} File name.}

\item{skip}{\code{numeric(1)} If not zero, will skip however many desired lines in the
file before starting to read.}
}
\value{
\code{list} \linkS4class{universalmotif} objects.
}
\description{
Import HOMER formatted motifs. See \url{http://homer.ucsd.edu/homer/motif/}.
Assumed to be DNA motifs.
}
\examples{
homer <- read_homer(system.file("extdata", "homer.txt",
                                package = "universalmotif"))

}
\references{
\insertRef{homer}{universalmotif}
}
\seealso{
Other read_motifs: \code{\link{read_cisbp}},
  \code{\link{read_jaspar}}, \code{\link{read_matrix}},
  \code{\link{read_meme}}, \code{\link{read_motifs}},
  \code{\link{read_transfac}}, \code{\link{read_uniprobe}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{read_motifs}
