% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jaspar.R
\name{read_jaspar}
\alias{read_jaspar}
\title{[UNDER CONSTRUCTION] Load Jaspar motifs from a text file.}
\usage{
read_jaspar(motif_file, verbose = FALSE, mot_length_cutoff = NULL)
}
\arguments{
\item{motif_file}{Character. Text file containing motifs.}

\item{verbose}{Logical.}

\item{mot_length_cutoff}{Integer.}
}
\value{
A list of motifs.
}
\description{
Motifs can be a mix of raw and Jaspar formatted motifs, as described
in \url{http://jaspar.genereg.net/html/TEMPLATES/help.html}.
}
\examples{
   motifs <- system.file("extdata", "example.jaspar",
                         package = "universalmotif")
   rmotifs <- read_jaspar(motifs)

}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
