% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jaspar.R
\name{read_jaspar}
\alias{read_jaspar}
\title{Import JASPAR motifs.}
\usage{
read_jaspar(file, skip = 0)
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}
}
\description{
Import JASPAR motifs.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
