% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matrix.R
\name{read_matrix}
\alias{read_matrix}
\title{Import motifs from raw matrices.}
\usage{
read_matrix(file, skip = 0, type, positions = "columns", alphabet = "DNA",
  sep = "", headers = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{type}{Character. One of 'PCM', 'PPM', 'PWM', and 'ICM'. If missing will
try and guess which one.}

\item{positions}{Character. One of 'columns' or 'rows'.}

\item{alphabet}{Character. 'DNA', 'RNA', 'AA', or a string of letters.}

\item{sep}{Character. Indicates how individual motifs are seperated.}

\item{headers}{Logical or character, indicating if and how to read names.}

\item{BPPARAM}{Param for bplapply.}
}
\value{
List of universalmotif objects.
}
\description{
Import motifs from raw matrices.
}
\examples{
   hocomoco <- system.file("extdata", "hocomoco.txt", package = "universalmotif")
   hocomoco <- read_matrix(hocomoco, headers = ">", positions = "rows")

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
