% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meme.R
\name{read_meme}
\alias{read_meme}
\title{Import MEME motifs.}
\usage{
read_meme(file, skip = 0, readsites = FALSE, BPPARAM = SerialParam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{readsites}{Logical. If \code{TRUE}, the motif sites will be read
as well.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
List of universalmotif objects. If \code{readsites = TRUE}, a list
        comprising of a sub-list of motif objects and a sub-list of 
        motif sites will be returned.
}
\description{
Import MEME motifs.
}
\examples{
meme.minimal <- read_meme(system.file("extdata", "meme_minimal.txt",
                                      package = "universalmotif"))
meme.full <- read_meme(system.file("extdata", "meme_full.txt",
                                   package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
