% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meme.R
\name{read_meme}
\alias{read_meme}
\title{[UNDER CONSTRUCTION] Load MEME motifs from a text file.}
\usage{
read_meme(motif_file, verbose = FALSE, show_warnings = TRUE,
  use_alt_title = FALSE, mot_length_cutoff = NULL,
  source_sites_cutoff = NULL, e_val_cutoff = NULL, out_class = "matrix-2",
  motif_type = "lpm")
}
\arguments{
\item{motif_file}{Character. Text file containing MEME format motifs.}

\item{verbose}{Logical. \code{read_meme} will return information regarding any 
detected information in the file preamble.}

\item{show_warnings}{Logical. Set this option to FALSE if you using this function
in a loop and are sure the warnings can be ignored.}

\item{use_alt_title}{Logical. Set this to TRUE if you wish to use the
alternate name in the resulting motif object.}

\item{mot_length_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are shorter than specified.}

\item{source_sites_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are found in fewer numbers than
requested.}

\item{e_val_cutoff}{Double. \code{read_meme} will skip over motifs with an
E value higher than specified.}

\item{out_class}{Character. Specify the class the resulting motif objects
  will be stored as. These classes include:
\itemize{
  \item \code{matrix-1} Matrix with alphabet letters as rows.
  \item \code{matrix-2} Matrix with alphabet letters as columns.
}}

\item{motif_type}{Character. MEME motifs can be provided as either letter-
probability matrices or log-odds matrices. The load the former, set this
option to "lpm"; for the latter, "lom".}
}
\value{
A list of motif objects of the specified class.
}
\description{
\code{read_meme} can load MEME motifs from a file connection to a list of
motifs in the class specified by the user. Both full MEME format and 
minimal MEME format files can be read; the latter requires at least a MEME
version in the preamble. \code{read_meme} was written for MEME version 4 and
above; reading a file with a lower version of MEME will throw a warning, as
this may result in incorrect parsing.
}
\examples{
  motifs <- system.file("extdata", "minimal.meme", package = "universalmotif")
  rmotifs <- read_meme(motifs, show_warnings = FALSE, out_class = "matrix-2")

}
\author{
Benjamin Tremblay <b2trembl@uwaterloo.ca>
}
