% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meme.R
\name{read_meme}
\alias{read_meme}
\title{Import MEME motifs.}
\usage{
read_meme(file, skip = 0, BPPARAM = bpparam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{BPPARAM}{Param for bplapply.}
}
\value{
List of universalmotif objects.
}
\description{
Import MEME motifs.
}
\examples{
meme.minimal <- read_meme(system.file("extdata", "meme_minimal.txt",
                                      package = "universalmotif"))
meme.full <- read_meme(system.file("extdata", "meme_full.txt",
                                   package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
