% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meme.R, R/read_memeOLD.R
\name{read_meme}
\alias{read_meme}
\alias{read_meme}
\title{[UNDER CONSTRUCTION] Load MEME motifs from a text file.}
\usage{
read_meme(motif_file, verbose = FALSE, show_warnings = TRUE,
  use_alt_title = FALSE, mot_length_cutoff = NULL,
  source_sites_cutoff = NULL, e_val_cutoff = NULL, out_class = "by-col",
  motif_type = "lpm")

read_meme(motif_file, verbose = FALSE, show_warnings = TRUE,
  use_alt_title = FALSE, mot_length_cutoff = NULL,
  source_sites_cutoff = NULL, e_val_cutoff = NULL, out_class = "by-col",
  motif_type = "lpm")
}
\arguments{
\item{motif_file}{Character. Text file containing MEME format motifs.}

\item{verbose}{Logical. \code{read_meme} will return information regarding any 
detected information in the file preamble.}

\item{show_warnings}{Logical. Set this option to FALSE if you using this function
in a loop and are sure the warnings can be ignored.}

\item{use_alt_title}{Logical. Set this to TRUE if you wish to use the
alternate title in the resulting motif object.}

\item{mot_length_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are shorter than specified.}

\item{source_sites_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are found in fewer numbers than
requested.}

\item{e_val_cutoff}{Double. \code{read_meme} will skip over motifs with an
E value higher than specified.}

\item{out_class}{Character. Specify the class the resulting motif objects
will be stored as.}

\item{motif_type}{Character. MEME motifs can be provided as either PWM or
PFM. This option will specify which is loaded into R.}

\item{motif_file}{Character. Text file containing MEME format motifs.}

\item{verbose}{Logical. \code{read_meme} will return information regarding any 
detected information in the file preamble.}

\item{show_warnings}{Logical. Set this option to FALSE if you using this function
in a loop and are sure the warnings can be ignored.}

\item{use_alt_title}{Logical. Set this to TRUE if you wish to use the
alternate title in the resulting motif object.}

\item{mot_length_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are shorter than specified.}

\item{source_sites_cutoff}{Integer. Setting this option will cause \code{
read_meme} to skip over motifs which are found in fewer numbers than
requested.}

\item{e_val_cutoff}{Double. \code{read_meme} will skip over motifs with an
E value higher than specified.}

\item{out_class}{Character. Specify the class the resulting motif objects
will be stored as.}

\item{motif_type}{Character. MEME motifs can be provided as either PWM or
PFM. This option will specify which is loaded into R.}
}
\value{
A list of motif objects of the specified class.

A list of motif objects of the specified class.
}
\description{
\code{read_meme} can load MEME motifs from a file connection to a list of
motifs in the class specified by the user. Both full MEME format and 
minimal MEME format files can be read; the latter requires at least a MEME
version in the preamble. \code{read_meme} was written for MEME version 4 and
above; reading a file with a lower version of MEME will throw an error, as
this may result in incorrect parsing.

\code{read_meme} can load MEME motifs from a file connection to a list of
motifs in the class specified by the user.
}
\examples{
\dontrun{read_meme("motifs.txt", out_class = "seqLogo-pwm")  # not run.}
\dontrun{read_meme("motifs.txt", out_class = "seqLogo-pwm")}
}
