% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meme.R
\name{read_meme}
\alias{read_meme}
\title{Import MEME motifs.}
\usage{
read_meme(file, skip = 0, readsites = FALSE)
}
\arguments{
\item{file}{\code{character(1)} File name.}

\item{skip}{\code{numeric(1)} If not zero, will skip however many desired lines in the
file before starting to read.}

\item{readsites}{\code{logical(1)} If \code{TRUE}, the motif sites will be read as well.}
}
\value{
\code{list} \linkS4class{universalmotif} objects. If \code{readsites = TRUE}, a list
comprising of a sub-list of motif objects and a sub-list of
motif sites will be returned.
}
\description{
Import MEME formatted motifs, as well as original motif sequences. See
\url{http://meme-suite.org/}. Both 'full' and 'minimal' formats are
supported.
}
\examples{
meme.minimal <- read_meme(system.file("extdata", "meme_minimal.txt",
                                      package = "universalmotif"))
meme.full <- read_meme(system.file("extdata", "meme_full.txt",
                                   package = "universalmotif"))

}
\references{
\insertRef{meme}{universalmotif}
}
\seealso{
Other read_motifs: \code{\link{read_cisbp}},
  \code{\link{read_homer}}, \code{\link{read_jaspar}},
  \code{\link{read_matrix}}, \code{\link{read_transfac}},
  \code{\link{read_uniprobe}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{read_motifs}
