% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_motif.R
\name{read_motifs}
\alias{read_motifs}
\title{[UNDER CONSTRUCTION] Load motifs from a text file.}
\usage{
read_motifs(motif_file, verbose = FALSE, motif_format = "autodetect",
  out_class = "matrix-2", ...)
}
\arguments{
\item{motif_file}{Character.}

\item{verbose}{Logical.}

\item{motif_format}{Character.}

\item{out_class}{Character.}

\item{...}{Additional arguments to be passed to the motif format-specific
function.}
}
\value{
A list of motifs.
}
\description{
Motifs can be autodetected or explicitly named. Each file must only contain
one format; mixed motif files are not supported.
}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
