% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transfac.R
\name{read_transfac}
\alias{read_transfac}
\title{Import TRANSFAC motifs.}
\usage{
read_transfac(file, skip = 0)
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}
}
\description{
Assumed to be DNA motifs, type PCM.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
