% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transfac.R
\name{read_transfac}
\alias{read_transfac}
\title{Import TRANSFAC motifs.}
\usage{
read_transfac(file, skip = 0)
}
\arguments{
\item{file}{\code{character(1)} File name.}

\item{skip}{\code{numeric(1)} If not zero, will skip however many desired lines in the
file before starting to read.}
}
\value{
\code{list} [universalmotif-class objects.
}
\description{
Import TRANSFAC formatted motifs. Assumed to be DNA motifs, type PCM.
See \code{system.file("extdata", "transfac.txt", pacakge="universalmotif")}
for an example motif.
}
\examples{
transfac <- read_transfac(system.file("extdata", "transfac.txt",
                                      package = "universalmotif"))

}
\references{
\insertRef{transfac}{universalmotif}
}
\seealso{
Other read_motifs: \code{\link{read_cisbp}},
  \code{\link{read_homer}}, \code{\link{read_jaspar}},
  \code{\link{read_matrix}}, \code{\link{read_meme}},
  \code{\link{read_uniprobe}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{read_motifs}
