% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transfac.R
\name{read_transfac}
\alias{read_transfac}
\title{Import TRANSFAC motifs.}
\usage{
read_transfac(file, skip = 0, BPPARAM = bpparam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{BPPARAM}{Param for bplapply.}
}
\value{
List of universalmotif objects.
}
\description{
Assumed to be DNA motifs, type PCM.
}
\examples{
transfac <- read_transfac(system.file("extdata", "transfac.txt",
                                      package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
