% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uniprobe.R
\name{read_uniprobe}
\alias{read_uniprobe}
\title{Import UNIPROBE motifs.}
\usage{
read_uniprobe(file, skip = 0)
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}
}
\description{
Assumed DNA.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
