% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uniprobe.R
\name{read_uniprobe}
\alias{read_uniprobe}
\title{[UNDER CONSTRUCTION] Load UNIPROBE motifs from a text file.}
\usage{
read_uniprobe(motif_file, verbose = FALSE, mot_length_cutoff = NULL,
  out_class = "matrix-2")
}
\arguments{
\item{motif_file}{Character.}

\item{verbose}{Logical.}

\item{mot_length_cutoff}{Integer.}

\item{out_class}{Character.}
}
\value{
a list of motif objects of the desired class.
}
\description{
Support for uniprobe-style motifs. Each motif has a header and a number
matrix. Only DNA.
}
\examples{
   motifs <- system.file("extdata", "uniprobe.txt",
                         package = "universalmotif")
   rmotifs <- read_uniprobe(motifs)

}
\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
