% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uniprobe.R
\name{read_uniprobe}
\alias{read_uniprobe}
\title{Import UNIPROBE motifs.}
\usage{
read_uniprobe(file, skip = 0, BPPARAM = SerialParam())
}
\arguments{
\item{file}{Character.}

\item{skip}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
List of universalmotif objects.
}
\description{
Assumed DNA.
}
\examples{
uniprobe.minimal <- read_uniprobe(system.file("extdata", "uniprobe_minimal.txt",
                                              package = "universalmotif"))
uniprobe.full <- read_uniprobe(system.file("extdata", "uniprobe_full.txt",
                                           package = "universalmotif"))

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
