% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cbind}
\alias{subset}
\alias{rownames}
\alias{colnames}
\alias{ncol}
\alias{nrow}
\alias{rowSums}
\alias{colSums}
\alias{colMeans}
\alias{rowMeans}
\alias{normalize}
\alias{as.data.frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}, \code{\link[BiocGenerics]{cbind}}, \code{\link[BiocGenerics:colSums]{colMeans}}, \code{\link[BiocGenerics:row_colnames]{colnames}}, \code{\link[BiocGenerics:colSums]{colSums}}, \code{\link[BiocGenerics:nrow]{ncol}}, \code{\link[BiocGenerics]{normalize}}, \code{\link[BiocGenerics]{nrow}}, \code{\link[BiocGenerics:colSums]{rowMeans}}, \code{\link[BiocGenerics:row_colnames]{rownames}}, \code{\link[BiocGenerics:colSums]{rowSums}}, \code{\link[BiocGenerics]{subset}}}
}}

