% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sites.R
\name{sample_sites}
\alias{sample_sites}
\title{Generate binding sites from a motif.}
\usage{
sample_sites(motif, n = 100, BPPARAM = SerialParam())
}
\arguments{
\item{motif}{Motif object.}

\item{n}{Numeric. Number of sites to generate.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
\linkS4class{XStringSet} object.
}
\description{
Given probabilities for a sequence as represented by a motif, generate
random sequences with the same length as the motif.
}
\examples{
motif <- create_motif()
sites <- sample_sites(motif)

}
\seealso{
\code{\link{create_sequences}}, \code{\link{create_motif}},
   \code{\link{add_multifreq}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
