% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_sequences.R
\name{scan_sequences}
\alias{scan_sequences}
\title{Find motif binding sites in a set of sequences.}
\usage{
scan_sequences(motifs, sequences, threshold = 0.6, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a single motif.}

\item{sequences}{XStringSet object. List of sequences to scan}

\item{threshold}{Numeric. Logodds threshold.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Site search results as a data.frame object.
}
\description{
Find motif binding sites in a set of sequences.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
