% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_sequences.R
\name{scan_sequences}
\alias{scan_sequences}
\title{Scan sequences for matches to input motifs.}
\usage{
scan_sequences(motifs, sequences, threshold = 0.001,
  threshold.type = "pvalue", RC = FALSE, use.freq = 1, verbose = 1,
  progress = TRUE, BP = FALSE)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable motif formats.}

\item{sequences}{\code{XStringSet} Sequences to scan. Alphabet should
match motif.}

\item{threshold}{\code{numeric(1)} Between 0 and 1. See details.}

\item{threshold.type}{\code{character(1)} One of \code{c('logodds', 'pvalue')}.
See details.}

\item{RC}{\code{logical(1)} If \code{TRUE}, check reverse complement of input
sequences.}

\item{use.freq}{\code{numeric(1)} The default, 1, uses the motif matrix (from
the \code{motif["motif"]} slot) to search for sequences. If a higher
number is used, then the matching k-let matrix from the
\code{motif["multifreq"]} slot is used. See \code{\link{add_multifreq}}.}

\item{verbose}{\code{numeric(1)} Describe progress, from none (0) to very
verbose (3).}

\item{progress}{\code{logical(1)} Show progress.}

\item{BP}{\code{logical(1)} Use BiocParallel.}
}
\value{
\code{data.frame} with each row representing one hit; if the input
   sequences are \code{DNAStringSet} or \code{RNAStringSet}, then an
   additional column with the strand is included.
}
\description{
For sequences of any alphabet, scan them using the PWM matrices of
a set of input motifs.
}
\details{
Similar to \code{\link[Biostrings]{matchPWM}}, the scanning method uses
   logodds scoring. (To see the scoring matrix for any motif, simply
   run \code{convert_type(motif, "PWM")}; for a \code{multifreq} scoring
   matrix: \code{apply(motif["multifreq"]$`2`, 2, ppm_to_pwm)}). In order
   to score a sequence, at each position within a sequence of length equal
   to the length of the motif, the scores for each base are summed. If the
   score sum is above the desired threshold, it is kept.

   If \code{threshold.type = 'logodds'}, then to calculate the minimum
   allowed score the maximum possible score for a motif is multiplied
   by the value set by \code{threshold}. To determine the maximum
   possible score a motif (of type PWM), run
   \code{sum(apply(motif['motif'], 2, max))}.

   If \code{threshold.type = 'pvalue'}, then threshold logodds scores are
   generated using \code{\link{motif_pvalue}}.

   Note: memory usage increases exponentially with k.
}
\examples{
## any alphabet can be used
\dontrun{
set.seed(1)
alphabet <- paste(c(letters), collapse = "")
motif <- create_motif("hello", alphabet = alphabet)
sequences <- create_sequences(alphabet, seqnum = 1000, seqlen = 100000)
scan_sequences(motif, sequences)
}

}
\references{
\insertRef{biostrings}{universalmotif}
}
\seealso{
\code{\link{add_multifreq}}, \code{\link[Biostrings]{matchPWM}},
   \code{\link{enrich_motifs}}, \code{\link{motif_pvalue}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
