% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_sequences.R
\name{scan_sequences}
\alias{scan_sequences}
\title{Find motif binding sites in a set of sequences.}
\usage{
scan_sequences(motifs, sequences, threshold = 0.001,
  threshold.type = "pvalue", RC = FALSE, use.freq = 1,
  verbose = TRUE, progress_bar = FALSE, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{\linkS4class{universalmotif} objects.}

\item{sequences}{XStringSet object. Sequences to scan. Alphabet should
match motif.}

\item{threshold}{Numeric. Between 0 and 1. See details.}

\item{threshold.type}{Character. One of 'logodds' and 'pvalue'. See details.}

\item{RC}{Logical. If \code{TRUE}, check reverse complement of input
sequences.}

\item{use.freq}{Numeric. The default, 1, uses the motif matrix (from
the \code{motif["motif"]} slot) to search for sequences. If a higher
number is used, then the matching k-let matrix from the
\code{motif["multifreq"]} slot is used. See \code{\link{add_multifreq}}.}

\item{progress_bar}{Logical. Shoe progress bar.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Results as a data.frame object.
}
\description{
Find matches to motifs in a set of input sequences.
}
\details{
If \code{use.freq = 1} and the sequences are DNAStringSet or RNAStringSet
   class, then the \code{\link[Biostrings]{matchPWM}} function from the
   Biostrings package is used. Otherwise, a different (and less efficient)
   scanning function is used.
   
   Similar to \code{\link[Biostrings]{matchPWM}}, the scanning method uses
   logodds scoring. (To see the scoring matrix for any motif, simply
   run \code{convert_type(motif, "PWM")}; for a \code{multifreq} scoring
   matrix: \code{apply(motif["multifreq"]$`2`, 2, ppm_to_pwm)}). In order
   to score a sequence, at each position within a sequence of length equal
   to the length of the motif, the scores for each base are summed. If the
   score sum is above the desired threshold, it is kept.

   If \code{threshold.type = 'logodds'}, then to calculate the minimum
   allowed score the maximum possible score for a motif is multiplied
   by the value set by \code{threshold}. To determine the maximum
   possible score a motif (of type PWM), run
   \code{sum(apply(motif['motif'], 2, max))}.

   Note: memory usage increases exponentially with k.
}
\examples{
# any alphabet can be used
\dontrun{
set.seed(1)
alphabet <- paste(c(letters), collapse = "")
motif <- create_motif("hello", alphabet = alphabet)
sequences <- create_sequences(alphabet, seqnum = 1000, seqlen = 100000)
scan_sequences(motif, sequences)
}

}
\references{
\insertRef{biostrings}{universalmotif}

   \insertRef{tfmpvalue}{universalmotif}
}
\seealso{
\code{\link{add_multifreq}}, \code{\link[Biostrings]{matchPWM}},
   \code{\link{enrich_motifs}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
