% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_sequences.R
\name{scan_sequences}
\alias{scan_sequences}
\title{Find motif binding sites in a set of sequences.}
\usage{
scan_sequences(motifs, sequences, threshold = 0.6, RC = FALSE,
  HMMorder = 0, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a single motif.}

\item{sequences}{XStringSet object. List of sequences to scan}

\item{threshold}{Numeric. Logodds threshold.}

\item{RC}{Logical. Check reverse strand.}

\item{HMMorder}{Numeric.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Site search results as a data.frame object.
}
\description{
Find motif binding sites in a set of sequences.
}
\details{
Benchmarking: scanning a 1 billion bp sequence with HMMorder = 1 and 
   RC = FALSE took about 4.5 minutes. Scanning 1000 sequences 1 million bp
   each took about 6 minutes.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
