% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_motifs.R
\name{shuffle_motifs}
\alias{shuffle_motifs}
\title{Shuffle motifs by column.}
\usage{
shuffle_motifs(motifs, k = 2, method = "linear", leftovers = "asis")
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable formats.}

\item{k}{\code{numeric(1)} K-let size.}

\item{method}{\code{character(1)} One of \code{c('linear', 'random')}.
Only relevant if \code{k > 1}. See details.}

\item{leftovers}{\code{character(1)} For \code{method = 'random'}. One of
\code{c('asis', 'first', 'split', 'discard')}. See details.}
}
\value{
Motifs. See \code{\link[=convert_motifs]{convert_motifs()}} for available output
formats.
}
\description{
Given a set of motifs, shuffle the columns between them. Currently does not
support keeping the 'multifreq' slot. Only the 'bkg', 'nsites', 'strand',
and 'bkgsites' slots will be preserved. Uses the same shuffling methods
as \code{\link[=shuffle_sequences]{shuffle_sequences()}}. When shuffling more than one motif, they are
shuffled together.
}
\details{
If \code{method = 'linear'}, then the input positions are split linearly every
\code{k} columns after which they are shuffled. If \code{method = random}, then
sets of \code{k}-columns are chosen randonmly before being shuffled. This leaves
leftover column islands smaller than \code{k}; these can be left \code{asis}, placed
\code{first}, \code{split} between the beginning and the end, or \code{discard}ed. See
\code{\link[=shuffle_motifs]{shuffle_motifs()}}.
}
\seealso{
\code{\link[=shuffle_sequences]{shuffle_sequences()}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
