% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_motifs.R
\name{shuffle_motifs}
\alias{shuffle_motifs}
\title{Shuffle motifs by column.}
\usage{
shuffle_motifs(motifs, shuffle.k = 2, shuffle.method = "linear",
  shuffle.leftovers = "asis", BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{Motifs.}

\item{shuffle.k}{Numeric.}

\item{shuffle.method}{Character.}

\item{shuffle.leftovers}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\description{
Given a set of motifs, shuffle the columns between them. Currently does not
support keeping the 'multifreq' slot. Only the 'bkg', 'nsites', 'strand',
and 'bkgsites' slots will be preserved.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
