% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_motifs.R
\name{shuffle_motifs}
\alias{shuffle_motifs}
\title{Shuffle motifs by column.}
\usage{
shuffle_motifs(motifs, k = 2, method = "linear", leftovers = "asis",
  BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable formats.}

\item{k}{\code{numeric(1)} K-let size.}

\item{method}{\code{character(1)} One of \code{c('linear', 'random')}.
See details.}

\item{leftovers}{\code{character(1)} For \code{method = 'random'}. One of
\code{c('asis', 'first', 'split', 'discard')}. See details.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Motifs. See \code{\link{convert_motifs}} for available output
   formats.
}
\description{
Given a set of motifs, shuffle the columns between them. Currently does not
support keeping the 'multifreq' slot. Only the 'bkg', 'nsites', 'strand',
and 'bkgsites' slots will be preserved. Uses the same shuffling methods
as \code{\link{shuffle_sequences}}.
}
\seealso{
\code{\link{shuffle_sequences}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
