% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_sequences.R
\name{shuffle_sequences}
\alias{shuffle_sequences}
\title{Shuffle input sequences.}
\usage{
shuffle_sequences(sequences, k = 1, method = "linear", leftovers = "asis",
  BPPARAM = SerialParam())
}
\arguments{
\item{sequences}{XStringSet objects.}

\item{k}{Numeric. k-let size.}

\item{method}{Character. One of 'markov', 'linear', and 'random'.}

\item{leftovers}{Character. For \code{method = 'random'}.}

\item{BPPARAM}{See \code{\link[BiocParallel]{SerialParam}}.}
}
\value{
XStringSet object.
}
\description{
Shuffle input sequences.
}
\details{
If one of \code{keepDI} and \code{keepTRI} is \code{TRUE}, then
   the Markov model \insertCite{markovmodel}{universalmotif} is used to
   generate sequences which will maintain (on average) the di/trinucleotide
   frequencies. Please note that this method is not a 'true' shuffling, and
   for short sequences (e.g. <100bp) this can result in slightly more
   dissimilar sequences versus true shuffling. See
   \insertCite{markovmodel;textual}{universalmotif} and 
   \insertCite{markovmodel2;textual}{universalmotif} for a discussion on the
   topic.
}
\references{
\insertRef{markovmodel}{universalmotif}

   \insertRef{markovmodel2}{universalmotif}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
