% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_sequences.R
\name{shuffle_sequences}
\alias{shuffle_sequences}
\title{Shuffle input sequences.}
\usage{
shuffle_sequences(sequences, inter = FALSE, keepDI = FALSE,
  keepTRI = FALSE)
}
\arguments{
\item{sequences}{XStringSet objects.}

\item{inter}{Logical. Shuffle letters between sequences? If \code{FALSE},
then the alphabet frequencies for each sequence are maintained.}

\item{keepDI}{Logical. Keep dinucleotide frequencies. Only supported for
DNA and RNA.}

\item{keepTRI}{Logical. Keep trinucleotide frequencies. Only supported for
DNA and RNA.}
}
\value{
XStringSet object.
}
\description{
Shuffle input sequences.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
