% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_alph.R
\name{switch_alph}
\alias{switch_alph}
\title{Switch between DNA and RNA alphabets.}
\usage{
switch_alph(motifs)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable formats.}
}
\value{
The DNA/RNA version of the motifs.
}
\description{
Convert a motif from DNA to RNA, or RNA to DNA.
}
\examples{
DNA.motif <- create_motif()
RNA.motif <- switch_alph(DNA.motif)

}
\seealso{
\code{\link{create_motif}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
