% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif_df.R
\name{tidy-motifs}
\alias{tidy-motifs}
\alias{to_df}
\alias{update_motifs}
\alias{to_list}
\title{Tidy manipulation of motifs.}
\usage{
to_df(motifs, extrainfo = FALSE)

update_motifs(motif_df, extrainfo = FALSE)

to_list(motif_df, extrainfo = FALSE)
}
\arguments{
\item{motifs}{List of motifs.}

\item{extrainfo}{Use the \code{extrainfo} slot in the tidy \code{data.frame}. The
column names will be taken from the character vectors themselves, and
unnamed elements will be assigned a unique name. To add elements to the
slot, simply create new columns in the \code{data.frame}. Note that these will
be coerced into characters. If \code{extrainfo} is not set to \code{TRUE} in
\code{to_df()}, then the contents of the slot will be transferred to the
\code{data.frame}. If \code{extrainfo} is not set to \code{TRUE} in \code{update_motifs()}
or \code{to_list()}, then the extra columns will be discarded.}

\item{motif_df}{Motif \code{data.frame} generated by \code{\link[=to_df]{to_df()}}.}
}
\value{
For \code{\link[=to_df]{to_df()}}: a \code{data.frame} with the exposed slots as columns.

For \code{\link[=update_motifs]{update_motifs()}}: the updated \code{data.frame}.

For \code{\link[=to_list]{to_list()}}: a list of motifs.
}
\description{
Tidy manipulation of motifs.
}
\examples{
\dontrun{
library(universalmotif)
library(dplyr)

m <- c(create_motif(name = "motif A"), create_motif(name = "motif B"))

# Change the names of the motifs using the tidy way:
m <- m \%>\%
   to_df() \%>\%
   mutate(name = paste0(name, "-2")) \%>\%
   to_list()

# Add random metadata to be stord in the extrainfo slot:
m_df <- to_df(m)
m_df$MyMetadata <- c("Info_1", "Info_2")
m <- to_list(m_df, extrainfo = TRUE)
}

}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
