% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/universalmotif-methods.R,
%   R/methods.R
\docType{methods}
\name{trim_motifs}
\alias{trim_motifs}
\alias{trim_motifs,universalmotif-method}
\alias{trim_motifs,list-method}
\alias{trim_motifs,ANY-method}
\title{[UNDER CONSTRUCTION] Trim motifs.
Trim a list of motifs.}
\usage{
trim_motifs(motifs, ic_cutoff = 0.25)

\S4method{trim_motifs}{universalmotif}(motifs, ic_cutoff = 0.25)

\S4method{trim_motifs}{list}(motifs, ic_cutoff = 0.25)

\S4method{trim_motifs}{ANY}(motifs, ic_cutoff = 0.25)
}
\arguments{
\item{motifs}{List of motifs, or a single motif object.}

\item{min_ic}{Numeric. Information cutoff for trimming.}
}
\value{
List of motifs or single motif object.
}
\description{
[UNDER CONSTRUCTION] Trim motifs.
Trim a list of motifs.
}
\section{Methods (by class)}{
\itemize{
\item \code{universalmotif}: Trim \linkS4class{universalmotif}.

\item \code{list}: Trim a list of motifs.

\item \code{ANY}: Trim a list of motifs which are not
\linkS4class{universalmotif}.
}}

\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
