% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_motifs.R
\name{trim_motifs}
\alias{trim_motifs}
\title{Trim motifs.}
\usage{
trim_motifs(motifs, min.ic = 0.25)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable formats.}

\item{min.ic}{\code{numeric(1)} Minimum allowed information content.}
}
\value{
Motifs See \code{\link{convert_motifs}} for available output
   formats.
}
\description{
Remove edges of a motif with low information content.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.trimmed <- trim_motifs(jaspar)

}
\seealso{
\code{\link{create_motif}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
