% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_motifs.R
\name{trim_motifs}
\alias{trim_motifs}
\title{Trim motifs.}
\usage{
trim_motifs(motifs, IC_cutoff = 0.25, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{Motif object.}

\item{IC_cutoff}{Numeric. Minimum allowed information content.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
Motifs Motif object or list.
}
\description{
Remove edges of a motif with low information content.
}
\examples{
jaspar <- read_jaspar(system.file("extdata", "jaspar.txt",
                                  package = "universalmotif"))
jaspar.trimmed <- trim_motifs(jaspar)

}
\seealso{
\code{\link{create_motif}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
