% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R,
%   R/universalmotif-methods.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\alias{universalmotif}
\alias{[,universalmotif,ANY,ANY,ANY-method}
\alias{[,universalmotif-method}
\alias{[<-,universalmotif,ANY,ANY,ANY-method}
\alias{[<-,universalmotif-method}
\alias{universalmotif}
\alias{initialize,universalmotif-method}
\alias{show,universalmotif-method}
\title{universalmotif: Motif class.}
\usage{
\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i)

\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i) <- value

\S4method{initialize}{universalmotif}(.Object, name, altname, family, organism,
  motif, alphabet = "DNA", type, icscore, nsites, pseudoweight = 0.8, bkg,
  bkgsites, consensus, strand = "+-", pval, qval, eval, extrainfo)

\S4method{show}{universalmotif}(object)
}
\arguments{
\item{x}{universalmotif object.}

\item{i}{Slot.}

\item{value}{Object to replace slot with.}

\item{.Object}{universalmotif object.}

\item{name}{Character.}

\item{altname}{Character.}

\item{family}{Character.}

\item{organism}{Character.}

\item{motif}{Matrix.}

\item{alphabet}{Character.}

\item{type}{Character.}

\item{icscore}{Numeric.}

\item{nsites}{Numeric.}

\item{pseudoweight}{Numeric.}

\item{bkg}{Numeric.}

\item{bkgsites}{Numeric.}

\item{consensus}{Character.}

\item{strand}{Character.}

\item{pval}{Numeric.}

\item{qval}{Numeric.}

\item{eval}{Numeric.}

\item{extrainfo}{Character.}

\item{object}{universalmotif object.}
}
\value{
A motif object.
}
\description{
Container for motif objects.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. Motif name.}

\item{\code{altname}}{Character.}

\item{\code{family}}{Character.}

\item{\code{organism}}{Character.}

\item{\code{motif}}{Matrix. Contains motif.}

\item{\code{alphabet}}{Character. Can be: DNA, RNA, AA, custom.}

\item{\code{type}}{Character. Can be: PCM, PPM, PWM.}

\item{\code{icscore}}{Numeric. Total information content for all positions.}

\item{\code{nsites}}{Numeric. Total number of sites.}

\item{\code{pseudoweight}}{Numeric. Amount of smoothing to apply.}

\item{\code{bkg}}{Numeric. Background letter frequencies.}

\item{\code{bkgsites}}{Numeric.}

\item{\code{consensus}}{Character. Motif consensus sequence.}

\item{\code{strand}}{Character. '+' and/or '-'.}

\item{\code{pval}}{Numeric.}

\item{\code{qval}}{Numeric.}

\item{\code{eval}}{Numeric.}

\item{\code{extrainfo}}{List.}
}}

\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
