% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R,
%   R/universalmotif-methods.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\alias{universalmotif}
\alias{[,universalmotif,ANY,ANY,ANY-method}
\alias{[,universalmotif-method}
\alias{[<-,universalmotif,ANY,ANY,ANY-method}
\alias{[<-,universalmotif-method}
\alias{universalmotif}
\alias{initialize,universalmotif-method}
\alias{show,universalmotif-method}
\title{universalmotif: Motif class.}
\usage{
\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i)

\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i) <- value

\S4method{initialize}{universalmotif}(.Object, name, altname, family, organism,
  motif, alphabet = "DNA", type, icscore, nsites, pseudocount = 0.8, bkg,
  bkgsites, consensus, strand = "+-", pval, qval, eval, extrainfo)

\S4method{show}{universalmotif}(object)
}
\arguments{
\item{x}{universalmotif object.}

\item{i}{Slot.}

\item{value}{Object to replace slot with.}

\item{.Object}{universalmotif object.}

\item{name}{Character. Motif name.}

\item{altname}{Character. Alternate motif name.}

\item{family}{Character. Transcription factor family.}

\item{organism}{Character. Species of origin.}

\item{motif}{Matrix.}

\item{alphabet}{Character. 'DNA', 'RNA', 'AA', 'custom', or a combined 
string representing the letters.}

\item{type}{Character. 'PCM', 'PPM', 'PWM', or 'ICM'.}

\item{icscore}{Numeric. Total information content. Automatically generated.}

\item{nsites}{Numeric. Number of sites the motif was constructed from.}

\item{pseudocount}{Numeric. Correction to be applied to prevent \code{-Inf}
from apearing in PWM matrices.}

\item{bkg}{Numeric. Must sum to 1 and be equal in length to the alphabet
length.}

\item{bkgsites}{Numeric. Total number of sites used to find the motif.}

\item{consensus}{Character. Consensus string. Automatically generated for 
'DNA', 'RNA', and 'AA' alphabets.}

\item{strand}{Character. Whether the motif is specific to a certain strand.}

\item{pval}{Numeric. P-value associated with motif.}

\item{qval}{Numeric. Adjusted P-value associated with motif.}

\item{eval}{Numeric. E-value associated with motif.}

\item{extrainfo}{Character. Any other extra information, represented as
a named character vector.}

\item{object}{\linkS4class{universalmotif} object.}
}
\value{
A motif object of class \linkS4class{universalmotif}.
}
\description{
Container for motif objects. See \code{\link{create_motif}} for creating
motifs.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. Length 1.}

\item{\code{altname}}{Character. Length 0 or 1.}

\item{\code{family}}{Character. Length 0 or 1.}

\item{\code{organism}}{Character. Length 0 or 1.}

\item{\code{motif}}{Matrix.}

\item{\code{alphabet}}{Character. Length 1.}

\item{\code{type}}{Character. Length 1.}

\item{\code{icscore}}{Numeric. Length 1.}

\item{\code{nsites}}{Numeric. Length 0 or 1.}

\item{\code{pseudocount}}{Numeric. Length 1.}

\item{\code{bkg}}{Numeric. Length equal to number of letters in alphabet.}

\item{\code{bkgsites}}{Numeric. Lenght 0 or 1.}

\item{\code{consensus}}{Character. Length 0 or 1.}

\item{\code{strand}}{Character. Length 1.}

\item{\code{pval}}{Numeric. Length 0 or 1.}

\item{\code{qval}}{Numeric. Length 0 or 1.}

\item{\code{eval}}{Numeric. Length 0 or 1.}

\item{\code{extrainfo}}{Character. Length 0 or more.}
}}

\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
