% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\title{[UNDER CONSTRUCTION] universalmotif: Motif class.}
\description{
Placeholder description.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. Motif name.}

\item{\code{motif}}{Matrix. Contains motif.}

\item{\code{alphabet}}{Character. Describes alphabet. Can be: DNA, RNA, AA, custom.}

\item{\code{type}}{Character. Can be: PCM, PPM, or PWM.}

\item{\code{icscore}}{Numeric. Total information content for all positions.}

\item{\code{nsites}}{Numeric. Total number of sites containing motif.}

\item{\code{pseudoweight}}{Numeric. Amount of smoothing to apply.}

\item{\code{bkg}}{Numeric. Background letter frequencies.}

\item{\code{consensus}}{Character. Motif consensus sequence.}

\item{\code{strand}}{Character. '+' or '-'.`}

\item{\code{extra}}{Character.}
}}

\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
