% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R,
%   R/universalmotif-methods.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\alias{universalmotif}
\alias{[,universalmotif,ANY,ANY,ANY-method}
\alias{[,universalmotif-method}
\alias{[<-,universalmotif,ANY,ANY,ANY-method}
\alias{[<-,universalmotif-method}
\alias{initialize,universalmotif-method}
\alias{show,universalmotif-method}
\alias{as.data.frame,universalmotif-method}
\alias{subset,universalmotif-method}
\alias{normalize,universalmotif-method}
\alias{rowMeans,universalmotif-method}
\alias{colMeans,universalmotif-method}
\alias{colSums,universalmotif-method}
\alias{rowSums,universalmotif-method}
\alias{nrow,universalmotif-method}
\alias{ncol,universalmotif-method}
\alias{colnames,universalmotif-method}
\alias{rownames,universalmotif-method}
\alias{cbind,universalmotif-method}
\title{universalmotif: Motif class.}
\usage{
\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i)

\S4method{[}{universalmotif,ANY,ANY,ANY}(x, i) <- value

\S4method{initialize}{universalmotif}(.Object, name, altname, family,
  organism, motif, alphabet = "DNA", type, icscore, nsites,
  pseudocount = 0.8, bkg, bkgsites, consensus, strand = "+-", pval,
  qval, eval, multifreq, extrainfo)

\S4method{show}{universalmotif}(object)

\S4method{as.data.frame}{universalmotif}(x)

\S4method{subset}{universalmotif}(x, select)

\S4method{normalize}{universalmotif}(object)

\S4method{rowMeans}{universalmotif}(x)

\S4method{colMeans}{universalmotif}(x)

\S4method{colSums}{universalmotif}(x)

\S4method{rowSums}{universalmotif}(x)

\S4method{nrow}{universalmotif}(x)

\S4method{ncol}{universalmotif}(x)

\S4method{colnames}{universalmotif}(x)

\S4method{rownames}{universalmotif}(x)

\S4method{cbind}{universalmotif}(..., deparse.level = 1)
}
\arguments{
\item{x}{\linkS4class{universalmotif} Motif.}

\item{i}{\code{character} Slot.}

\item{value}{Object to replace slot with.}

\item{.Object}{\linkS4class{universalmotif} Final motif.}

\item{name}{\code{character} Motif name.}

\item{altname}{\code{character} Alternate motif name.}

\item{family}{\code{character} Transcription factor family.}

\item{organism}{\code{character} Species of origin.}

\item{motif}{\code{matrix} Each column represents a position in the motif.}

\item{alphabet}{\code{character(1)} One of \code{c('DNA', 'RNA', 'AA', 'custom')},
or a combined string representing the letters.}

\item{type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}.}

\item{icscore}{\code{numeric} Total information content. Automatically generated.}

\item{nsites}{\code{numeric} Number of sites the motif was constructed from.}

\item{pseudocount}{\code{pseudocount(1)} Correction to be applied to prevent \code{-Inf}
from apearing in PWM matrices.}

\item{bkg}{\code{numeric} Must sum to 1 and be equal in length to the alphabet
length. If missing, assumes a uniform background.}

\item{bkgsites}{\code{numeric} Total number of sites used to find the motif.}

\item{consensus}{\code{character} Consensus string. Automatically generated for 
'DNA', 'RNA', and 'AA' alphabets.}

\item{strand}{\code{character(1)} Whether the motif is specific to a certain strand.}

\item{pval}{\code{numeric} P-value associated with motif.}

\item{qval}{\code{numeric} Adjusted P-value associated with motif.}

\item{eval}{\code{numeric} E-value associated with motif.}

\item{multifreq}{\code{list} See \code{\link{add_multifreq}}.}

\item{extrainfo}{\code{character} Any other extra information, represented as
a named character vector.}

\item{object}{\linkS4class{universalmotif} Motif.}

\item{select}{\code{numeric} Columns to keep.}

\item{...}{\linkS4class{universalmotif} Motifs.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{object}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}

\item{x}{\linkS4class{universalmotif} Single motif.}
}
\value{
A motif object of class \linkS4class{universalmotif}.
}
\description{
Container for motif objects. See \code{\link{create_motif}} for creating
motifs.
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character(1)}}

\item{\code{altname}}{\code{character}}

\item{\code{family}}{\code{character}}

\item{\code{organism}}{\code{character}}

\item{\code{motif}}{\code{matrix}}

\item{\code{alphabet}}{\code{character(1)}}

\item{\code{type}}{\code{character(1)}}

\item{\code{icscore}}{\code{numeric(1)} Generated automatically.}

\item{\code{nsites}}{\code{numeric}}

\item{\code{pseudocount}}{\code{numeric(1)}}

\item{\code{bkg}}{\code{numeric} Length equal to number of letters in alphabet.}

\item{\code{bkgsites}}{\code{numeric}}

\item{\code{consensus}}{\code{character} Generated automatically.}

\item{\code{strand}}{\code{character(1)}}

\item{\code{pval}}{\code{numeric}}

\item{\code{qval}}{\code{numeric}}

\item{\code{eval}}{\code{numeric}}

\item{\code{multifreq}}{\code{list}}

\item{\code{extrainfo}}{\code{character}}
}}

\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
