% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif-class.R,
%   R/universalmotif-methods.R, R/methods.R
\docType{class}
\name{universalmotif-class}
\alias{universalmotif-class}
\alias{universalmotif}
\alias{show,universalmotif-method}
\alias{initialize,universalmotif-method}
\alias{motif_slots,universalmotif-method}
\alias{motif_slots<-,universalmotif-method}
\alias{seqLogo,universalmotif-method}
\title{[UNDER CONSTRUCTION] universalmotif: Motif class.}
\usage{
\S4method{show}{universalmotif}(object)

\S4method{initialize}{universalmotif}(.Object, name, motif, alphabet = "DNA",
  type = character(0), icscore = numeric(0), nsites = numeric(0),
  pseudoweight = 0.8, bkg = numeric(0), consensus = character(0),
  strand = c("+", "-"), extrachar = character(0), extranum = numeric(0),
  pval = numeric(0), eval = numeric(0), bkgsites = numeric(0))

\S4method{motif_slots}{universalmotif}(object, slots)

\S4method{motif_slots}{universalmotif}(object, slot) <- value

\S4method{seqLogo}{universalmotif}(pwm, ic.scale = TRUE, xaxis = TRUE,
  yaxis = TRUE, xfontsize = 15, yfontsize = 15)
}
\value{
A motif object.
}
\description{
Placeholder description.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method.

\item \code{initialize}: Initialize method.

\item \code{motif_slots}: Accessor function.

\item \code{motif_slots<-}: Replacement method.

\item \code{seqLogo}: Plot a \linkS4class{universalmotif} object with seqLogo.
}}

\section{Slots}{

\describe{
\item{\code{name}}{Character. Motif name.}

\item{\code{motif}}{Matrix. Contains motif.}

\item{\code{alphabet}}{Character. Describes alphabet. Can be: DNA, RNA, AA, custom.}

\item{\code{type}}{Character. Can be: PCM, PPM, or PWM.}

\item{\code{icscore}}{Numeric. Total information content for all positions.}

\item{\code{nsites}}{Numeric. Total number of sites containing motif.}

\item{\code{pseudoweight}}{Numeric. Amount of smoothing to apply.}

\item{\code{bkg}}{Numeric. Background letter frequencies.}

\item{\code{consensus}}{Character. Motif consensus sequence.}

\item{\code{strand}}{Character. '+' or '-'.}

\item{\code{pval}}{Numeric.}

\item{\code{eval}}{Numeric.}

\item{\code{extrachar}}{Character.}

\item{\code{extranum}}{Numeric.}

\item{\code{bkgsites}}{Numeric.}
}}

\author{
Benjamin Tremblay, \email{b2trembl@uwaterloo.ca}
}
