% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif.R
\docType{package}
\name{universalmotif-pkg}
\alias{universalmotif-pkg}
\title{universalmotif: Import, Modify and Export Motifs with R}
\description{
universalmotif: Import, Modify and Export Motifs with R
}
\details{
A collection of utility functions for working with motifs.
}
