% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif.R
\docType{package}
\name{universalmotif-pkg}
\alias{universalmotif-pkg}
\alias{universalmotif-pkg-package}
\title{universalmotif: importing and export motifs with R}
\description{
A collection of utility functions for working with motifs.
}
