% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalmotif.R
\docType{package}
\name{universalmotif-pkg}
\alias{universalmotif-pkg}
\alias{universalmotif-pkg-package}
\title{universalmotif: import and export motifs with R}
\description{
universalmotif: import and export motifs with R
}
\details{
A collection of utility functions for working with motifs.
}
