% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_motifs.R, R/utils.R
\name{make_DBscores}
\alias{make_DBscores}
\alias{utilities}
\alias{ppm_to_icm}
\alias{icm_to_ppm}
\alias{pcm_to_ppm}
\alias{ppm_to_pcm}
\alias{ppm_to_pwm}
\alias{pwm_to_ppm}
\alias{position_icscore}
\alias{get_consensus}
\alias{consensus_to_ppm}
\alias{consensus_to_ppmAA}
\alias{get_consensusAA}
\title{Utility functions.}
\usage{
make_DBscores(db.motifs, method = "Pearson", shuffle.db = TRUE,
  shuffle.k = 3, shuffle.method = "linear", shuffle.leftovers = "asis",
  rand.tries = 1000, min.overlap = 6, min.mean.ic = 0,
  progress_bar = TRUE, BPPARAM = SerialParam())

ppm_to_icm(position, bkg, schneider_correction = FALSE, nsites,
  relative_entropy = FALSE)

icm_to_ppm(position)

pcm_to_ppm(position, pseudocount = 0.8)

ppm_to_pcm(position, nsites = 100)

ppm_to_pwm(position, bkg, pseudocount = 0.8, nsites = 100, smooth = TRUE)

pwm_to_ppm(position, bkg)

position_icscore(position, bkg, type, pseudocount = 0.8, nsites = 100,
  relative_entropy = FALSE)

get_consensus(position, alphabet = "DNA", type = "PPM", pseudocount = 0.8)

consensus_to_ppm(letter)

consensus_to_ppmAA(letter)

get_consensusAA(position, type, pseudocount)
}
\arguments{
\item{db.motifs}{List. Database motifs.}

\item{method}{Character. One of Pearson, Euclidean, KL.}

\item{shuffle.k}{Numeric.}

\item{shuffle.method}{Character.}

\item{shuffle.leftovers}{Character.}

\item{min.overlap}{Numeric. Minimum required motif overlap.}

\item{progress_bar}{Logical. Show progress.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{position}{Numeric. A numeric vector representing the frequency or
probability for each alphabet letter at a specific position.}

\item{bkg}{Numeric. Should be the same length as the alphabet length.}

\item{schneider_correction}{Logical. Apply sample size correction.}

\item{nsites}{Numeric. Number of sites motif originated from.}

\item{relative_entropy}{Logical. Calculate information content as
relative entropy or Kullback-Leibler divergence.}

\item{pseudocount}{Numeric. Used to prevent zeroes in motif matrix.}

\item{smooth}{Logical. Apply pseudocount correction.}

\item{type}{Character. 'PCM', 'PPM', 'PWM' or 'ICM'.}

\item{alphabet}{Character. 'DNA' or 'RNA'.}

\item{letter}{Character. Any DNA, RNA, or AA IUPAC letter. Ambiguity letters
are accepted.}

\item{tries}{Numeric. Number of random motifs to use for calculating db scores.}
}
\value{
For \code{ppm_to_icm}, \code{icm_to_ppm}, \code{pcm_to_ppm},
   \code{ppm_to_pcm}, \code{ppm_to_pwm}, and \code{pwm_to_ppm}: a numeric
   vector with length equal to input numeric vector.

   For \code{consensus_to_ppm} and \code{consensus_to_ppmAA}: a numeric
   vector of length 4 and 20, respectively.

   For \code{position_icscore}: a numeric vector of length 1.

   For \code{get_consensus} and \code{get_consensusAA}: a character vector
   of length 1.

   For \code{make_DBscores}: a data.frame with score distributions for the
   input database.
}
\description{
Various small functions used for motif creation.
}
\seealso{
\code{\link{create_motif}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
