% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{utilities}
\alias{utilities}
\title{Utility functions.}
\description{
Utility functions have been split into two categories: those related to
motifs ?\code{utils-motif}, and those related to sequences ?\code{utils-sequence}.
}
\seealso{
\link{utils-motif}, \link{utils-sequence}
}
\author{
Benjamin Jean-Marie Tremblay, \email{benjamin.tremblay@uwaterloo.ca}
}
