% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_motifs.R
\name{view_motifs}
\alias{view_motifs}
\title{Plot motif logos.}
\usage{
view_motifs(motifs, compare.type = "PPM", plot.type = "ICM",
  method = "Pearson", tryRC = TRUE, min.overlap = 6, min.mean.ic = 0.5,
  relative_entropy = FALSE, BPPARAM = SerialParam(), ...)
}
\arguments{
\item{motifs}{Motifs. If more than one, motifs will be aligned with first
motif in the list.}

\item{compare.type}{Character.}

\item{plot.type}{Character.}

\item{method}{Character.}

\item{tryRC}{Logical.}

\item{min.overlap}{Numeric.}

\item{min.mean.ic}{Numeric.}

\item{relative_entropy}{Logical.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{...}{For geom_logo.}
}
\description{
Plot motif logos.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
