% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_motifs.R
\name{view_motifs}
\alias{view_motifs}
\title{Plot motif logos.}
\usage{
view_motifs(motifs, use.type = "ICM", method = "ALLR", tryRC = TRUE,
  min.overlap = 6, min.mean.ic = 0.25, relative_entropy = FALSE,
  normalise.scores = FALSE, min.position.ic = 0, score.strat = "sum",
  return.raw = FALSE, dedup.names = FALSE, ...)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{use.type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}.}

\item{method}{\code{character(1)} One of PCC, EUCL, SW, KL, ALLR, BHAT, HELL,
SEUCL, MAN, ALLR_LL, WEUCL, WPCC. See details.}

\item{tryRC}{\code{logical(1)} Try the reverse complement of the motifs as well,
report the best score.}

\item{min.overlap}{\code{numeric(1)} Minimum overlap required when aligning the
motifs. Setting this to a number higher then the width of the motifs
will not allow any overhangs. Can also be a number less than 1,
representing the minimum fraction that the motifs must overlap.}

\item{min.mean.ic}{\code{numeric(1)} Minimum mean information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{relative_entropy}{\code{logical(1)} Change the ICM calculation affecting
\code{min.position.ic} and \code{min.mean.ic}. See \code{\link[=convert_type]{convert_type()}}.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions, as well as alignments between motifs of similar length.
Similarity scores are multiplied by the ratio of
aligned positions to the total number of positions in the larger motif,
and the inverse for distance scores.}

\item{min.position.ic}{\code{numeric(1)} Minimum information content required between
individual alignment positions for it to be counted in the final alignment
score. It is recommended to use this together with \code{normalise.scores = TRUE},
as this will help punish scores resulting from only a fraction of an
alignment.}

\item{score.strat}{\code{character(1)} How to handle column scores calculated from
motif alignments. "sum": add up all scores. "a.mean": take the arithmetic
mean. "g.mean": take the geometric mean. "median": take the median.
"wa.mean", "wg.mean": weighted arithmetic/geometric mean. "fzt": Fisher
Z-transform. Weights are the
total information content shared between aligned columns.}

\item{return.raw}{\code{logical(1)} Instead of returning a plot, return the
aligned named matrices used to generate the plot. This can be useful
if you wish to use \code{\link[=view_motifs]{view_motifs()}} alignment capabilities for custom
plotting uses. Alignment is performed by adding empty columns to the
left or right of motifs to generate matrices of equal length.}

\item{dedup.names}{\code{logical(1)} Plotting motifs with duplicated names is
not allowed. Setting this to \code{TRUE} allows the names to be modified
for plotting.}

\item{...}{Additional options for \code{\link[ggseqlogo:geom_logo]{ggseqlogo::geom_logo()}}.}
}
\value{
A ggplot object. If \code{return.raw = TRUE}, a list.
}
\description{
Show sequence logo. If given a list of more than one motif, then the motifs
are aligned with the first in the list.
}
\details{
Since the \pkg{ggseqlogo} package can only plot individual
characters and not strings, plotting the \code{multifreq} slot is not
supported. See the \code{examples} section for plotting the \code{multifreq}
slot using the \pkg{Logolas} package.

See \code{\link[=compare_motifs]{compare_motifs()}} for more info on comparison parameters.

Note: \code{score.strat = "a.mean"} is NOT recommended, as \code{\link[=view_motifs]{view_motifs()}} will
not discriminate between two alignments with equal mean scores, even if one
alignment is longer than the other.
}
\examples{
## plotting multifreq motifs:
\dontrun{
  motif <- create_motif()
  motif <- add_multifreq(motif, sample_sites(motif))
  Logolas::logomaker(motif["multifreq"][["2"]], type = "Logo",
                     color_type = "per_symbol")
}

}
\references{
\insertRef{logolas}{universalmotif}

\insertRef{ggseqlogo}{universalmotif}
}
\seealso{
\code{\link[=compare_motifs]{compare_motifs()}}, \code{\link[=add_multifreq]{add_multifreq()}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
