% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_motifs.R
\name{view_motifs}
\alias{view_motifs}
\title{Plot motif logos.}
\usage{
view_motifs(motifs, use.type = "ICM", method = "MPCC", tryRC = TRUE,
  min.overlap = 6, min.mean.ic = 0.5, relative_entropy = FALSE,
  normalise.scores = FALSE, ...)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable motif formats.}

\item{use.type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}.}

\item{method}{\code{character(1)} One of \code{c('Pearson', 'Euclidean', 'KL')}.}

\item{tryRC}{\code{logical(1)} Check if motif reverse complement leads to a
better alignment.}

\item{min.overlap}{\code{numeric(1)} Minimum alignment overlap between
motifs. If \code{min.overlap < 1}, this represents the minimum fraction
between the two motifs during alignment.}

\item{min.mean.ic}{\code{numeric(1)} Minimum information content between the
two motifs for an alignment to be scored. This helps prevent scoring
alignments between low information content regions of two motifs.}

\item{relative_entropy}{\code{logical(1)} For ICM calculation. See
\code{\link[=convert_type]{convert_type()}}.}

\item{normalise.scores}{\code{logical(1)} Favour alignments which leave fewer
unaligned positions.}

\item{...}{Addtional options for \code{\link[ggseqlogo:geom_logo]{ggseqlogo::geom_logo()}}.}
}
\value{
A \link[ggplot2:ggplot-class]{ggplot2::ggplot} object.
}
\description{
Show sequence logo. If given a list of more than one motif, then the motifs
are aligned with the first in the list.
}
\details{
Since the \pkg{ggseqlogo} package can only plot individual
characters and not strings, plotting the \code{multifreq} slot is not
supported. See the \code{examples} section for plotting the \code{multifreq}
slot using the \pkg{Logolas} package.
}
\examples{
## plotting multifreq motifs:
if (requireNamespace("Logolas", quietly = TRUE)) {
  motif <- create_motif()
  motif <- add_multifreq(motif, sample_sites(motif))
  Logolas::logomaker(motif["multifreq"][[as.character(2)]], type = "Logo",
                     color_type = "per_symbol")
}

}
\references{
\insertRef{logolas}{universalmotif}

\insertRef{ggseqlogo}{universalmotif}
}
\seealso{
\code{\link[=compare_motifs]{compare_motifs()}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
