% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_motifs.R
\name{view_motifs}
\alias{view_motifs}
\title{Plot motif logos.}
\usage{
view_motifs(motifs, use.type = "ICM", method = "Pearson", tryRC = TRUE,
  min.overlap = 6, min.mean.ic = 0.5, relative_entropy = FALSE,
  BPPARAM = SerialParam(), ...)
}
\arguments{
\item{motifs}{Single motif or a list of motifs.}

\item{use.type}{Character.}

\item{method}{Character.}

\item{tryRC}{Logical.}

\item{min.overlap}{Numeric.}

\item{min.mean.ic}{Numeric.}

\item{relative_entropy}{Logical.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{...}{Addtional options for \code{\link[ggseqlogo]{geom_logo}}.}
}
\description{
Show sequence logo. If given a list of more than one motif, then the motifs
are aligned with the first in the list.
}
\examples{
## plotting multifreq motifs:
if (requireNamespace(Logolas, quietly = TRUE)) {
  motif <- create_motif()
  motif <- add_multifreq(motif, sample_sites(motif))
  logomaker(motif["multifreq"][[`2`]], type = "Logo",
            color_type = "per_symbol")
}

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
