% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_homer.R
\name{write_homer}
\alias{write_homer}
\title{Export motifs in HOMER format.}
\usage{
write_homer(motifs, file, logodds_threshold = 0.6, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}

\item{logodds_threshold}{Stringency required for HOMER to match a motif.
See \code{\link{scan_sequences}}.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
NULL, invisibly.
}
\description{
Convert motifs to HOMER format and write to file.
}
\examples{
motif <- create_motif()
write_homer(motif, tempfile())

}
\references{
\insertRef{homer}{universalmotif}
}
\seealso{
\code{\link{read_homer}}

Other write_motifs: \code{\link{write_jaspar}},
  \code{\link{write_matrix}}, \code{\link{write_meme}},
  \code{\link{write_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
