% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_homer.R
\name{write_homer}
\alias{write_homer}
\title{Export motifs in HOMER format.}
\usage{
write_homer(motifs, file, logodds_threshold = 0.6, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}

\item{logodds_threshold}{Stringency required for HOMER to match a motif.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs in HOMER format.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
