% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_homer.R
\name{write_homer}
\alias{write_homer}
\title{Export motifs in HOMER format.}
\usage{
write_homer(motifs, file, logodds_threshold = 0.6)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable formats.}

\item{file}{\code{character(1)} File name.}

\item{logodds_threshold}{\code{numeric} Stringency required for HOMER to match a motif.
See \code{\link[=scan_sequences]{scan_sequences()}}.}
}
\value{
NULL, invisibly.
}
\description{
Convert DNA motifs to HOMER format and write to file.
See \url{http://homer.ucsd.edu/homer/motif/}.
}
\examples{
motif <- create_motif()
write_homer(motif, tempfile())

}
\references{
\insertRef{homer}{universalmotif}
}
\seealso{
\code{\link[=read_homer]{read_homer()}}

Other write_motifs: \code{\link{write_jaspar}},
  \code{\link{write_matrix}}, \code{\link{write_meme}},
  \code{\link{write_motifs}}, \code{\link{write_transfac}}
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{write_motifs}
