% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jaspar.R
\name{write_jaspar}
\alias{write_jaspar}
\title{Export motifs in JASPAR format.}
\usage{
write_jaspar(motifs, file)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable formats.}

\item{file}{\code{character(1)} File name.}
}
\value{
NULL, invisibly.
}
\description{
Convert motifs to JASPAR format and write to file.
}
\examples{
transfac <- read_transfac(system.file("extdata", "transfac.txt",
                                    package = "universalmotif"))
write_jaspar(transfac, tempfile())

}
\references{
\insertRef{jaspar}{universalmotif}
}
\seealso{
\code{\link{read_jaspar}}

Other write_motifs: \code{\link{write_homer}},
  \code{\link{write_matrix}}, \code{\link{write_meme}},
  \code{\link{write_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{write_motifs}
