% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jaspar.R
\name{write_jaspar}
\alias{write_jaspar}
\title{Export motifs in JASPAR format.}
\usage{
write_jaspar(motifs, file, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs in JASPAR format.
}
\examples{
transfac <- read_transfac(system.file("extdata", "transfac.txt",
                                    package = "universalmotif"))
write_jaspar(transfac, tempfile())

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
