% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_matrix.R
\name{write_matrix}
\alias{write_matrix}
\title{Export motifs as raw matrices.}
\usage{
write_matrix(motifs, file, positions = "columns", rownames = FALSE,
  type, sep = "", headers = TRUE, BPPARAM = SerialParam())
}
\arguments{
\item{motifs}{List of motifs or a single motif object.}

\item{file}{Character.}

\item{positions}{Character. One of 'columns' or 'rows'.}

\item{rownames}{Logical. Include alphabet letters as rownames.}

\item{type}{Character. One of 'PCM', 'PPM', 'PWM', and 'ICM'. If missing
will use whatever type the motif is currently stored as.}

\item{sep}{Character. Indicates how to separate individual motifs.}

\item{headers}{Logical or character, indicating if and how to write names.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}
}
\value{
NULL, invisibly.
}
\description{
Write motifs as simple matrices with optional headers to file.
}
\examples{
motif <- create_motif()
write_matrix(motif, tempfile(), headers = ">")

}
\seealso{
\code{\link{read_matrix}}

Other write_motifs: \code{\link{write_homer}},
  \code{\link{write_jaspar}}, \code{\link{write_meme}},
  \code{\link{write_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{write_motifs}
