% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_matrix.R
\name{write_matrix}
\alias{write_matrix}
\title{Export motifs as raw matrices.}
\usage{
write_matrix(motifs, file, positions = "columns", rownames = FALSE, type,
  sep = "", headers = TRUE, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a single motif object.}

\item{file}{Character.}

\item{positions}{Character. One of 'columns' or 'rows'.}

\item{rownames}{Logical. Include alphabet letters as rownames.}

\item{type}{Character. One of 'PCM', 'PPM', 'PWM', and 'ICM'. If missing
will use whatever type the motif is currently stored as.}

\item{sep}{Character. Indicates how to separate individual motifs.}

\item{BPPARAM}{See \code{\link[BiocParallel]{bpparam}}.}

\item{header}{Logical or character, indicating if and how to write names.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs as raw matrices.
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
