% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_matrix.R
\name{write_matrix}
\alias{write_matrix}
\title{Export motifs as raw matrices.}
\usage{
write_matrix(motifs, file, positions = "columns", rownames = FALSE,
  type, sep = "", headers = TRUE)
}
\arguments{
\item{motifs}{See \code{\link{convert_motifs}} for acceptable formats.}

\item{file}{\code{character(1)} File name.}

\item{positions}{\code{character(1)} One of \code{c('columns', 'rows')}.}

\item{rownames}{\code{logical(1)} Include alphabet letters as rownames.}

\item{type}{\code{character(1)} One of \code{c('PCM', 'PPM', 'PWM', 'ICM')}. If missing
will use whatever type the motif is currently stored as.}

\item{sep}{\code{character(1)} Indicates how to separate individual motifs.}

\item{headers}{\code{logical(1)}, \code{character(1)} Indicating if and how to write names.}
}
\value{
NULL, invisibly.
}
\description{
Write motifs as simple matrices with optional headers to file.
}
\examples{
motif <- create_motif()
write_matrix(motif, tempfile(), headers = ">")

}
\seealso{
\code{\link{read_matrix}}

Other write_motifs: \code{\link{write_homer}},
  \code{\link{write_jaspar}}, \code{\link{write_meme}},
  \code{\link{write_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{write_motifs}
