% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_meme}
\alias{write_meme}
\title{Create text files from motif list as MEME.}
\usage{
write_meme(motif_list, file_out, version = 4, bkg = c(0.25, 0.25, 0.25,
  0.25), strands = "+ -", alphabet = NULL)
}
\description{
Create text files from motif list as MEME.
}
