% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_meme.R
\name{write_meme}
\alias{write_meme}
\title{Export motifs in MEME format.}
\usage{
write_meme(motifs, file, version = 4, bkg, strand, BPPARAM = bpparam())
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}

\item{version}{MEME version.}

\item{bkg}{Background letter frequencies. If missing, will use background
frequencies from motif objects (if they are identical); else
background frequencies will be set to freq = 1/length(alphabet)}

\item{strand}{If missing, will use strand from motif objects (if identical);
otherwise will default to "+ -"}

\item{BPPARAM}{Param for bplapply.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs in MEME format.
}
\examples{
transfac <- read_transfac(system.file("extdata", "transfac.txt",
                                    package = "universalmotif"))
write_meme(transfac, tempfile())

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
