% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_motifs.R
\name{write_motifs}
\alias{write_motifs}
\title{Export motifs in universalmotif format.}
\usage{
write_motifs(motifs, file, minimal = FALSE, multifreq = TRUE)
}
\arguments{
\item{motifs}{See \code{\link[=convert_motifs]{convert_motifs()}} for acceptable formats.}

\item{file}{\code{character(1)} File name.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Write motifs as universalmotif objects to file.
}
\seealso{
Other write_motifs: \code{\link{write_homer}},
  \code{\link{write_jaspar}}, \code{\link{write_matrix}},
  \code{\link{write_meme}}, \code{\link{write_transfac}}
}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
\concept{write_motifs}
