% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transfac.R
\name{write_transfac}
\alias{write_transfac}
\title{Export motifs in TRANSFAC format.}
\usage{
write_transfac(motifs, file)
}
\arguments{
\item{motifs}{List of motifs or a motif object.}

\item{file}{Character.}
}
\value{
NULL, invisibly.
}
\description{
Export motifs in TRANSFAC format.
}
\examples{
\dontrun{
jaspar <- read_transfac(system.file("extdata", "jaspar.txt",
                                    package = "universalmotif"))
write_transfac(jaspar, "motifs.transfac")
}

}
\author{
Benjamin Tremblay, \email{b2tremblay@uwaterloo.ca}
}
